### 2020-11-12 #409 @vbaliga
pathviewR: Tools to import, clean, and visualize animal movement data in R
Submitting Author: Name (@vbaliga)
Repository:  https://github.com/vbaliga/pathviewr
Version submitted:   v0.9.4 (https://github.com/vbaliga/pathviewr/commit/6e04870d7451b05370255a6d0c7d6532f4cc1e5a)
Version accepted: TBD

---

```
Package: pathviewR
Title: Wrangle, Analyze, and Visualize Animal Movement Data
Version: 0.9.4
    c(person(given = "Melissa S.",
             family = "Armstrong",
             role = "aut",
             email = "melissa.armstrong@gmail.com",
      person(given = "Vikram B.",
             family = "Baliga",
             email = "vbaliga87@gmail.com",
      person(given = "Eric R.",
             family = "Press",
             role = "aut",
             email = "epress12@gmail.com",
      )
Description: Tools to import, clean, and visualize
    animal movement data from motion capture systems such as Optitrack's
    Motive, the Straw Lab's Flydra, or from other sources. We provide
    functions to remove artifacts, standardize tunnel position and tunnel
    axes, select a region of interest, isolate specific trajectories, fill
    gaps in trajectory data, and calculate 3D and per-axis velocity. For
    experiments of visual guidance, we also provide functions that use
    animal position to estimate perception of visual stimuli.
RoxygenNote: 7.1.1.9000
    R.matlab,
    data.table (>= 1.12.2),
    magrittr (>= 1.5),
    stringr (>= 1.4.0),
    tibble (>= 3.0.1),
    tidyr (>= 1.1.0),
    fANCOVA,
    purrr (>= 0.3.3),
    ggplot2 (>= 3.3.0),
    tidyselect (>= 1.1.0),
    cowplot
    covr

```

`pathviewR` provides tools to extract data from motion capture software such as [Optitrack's Motive](https://www.optitrack.com/software/motive/) and [the Straw Lab's Flydra](https://github.com/strawlab/flydra) (among other sources). Tools to wrangle and clean data sets to isolate specific trajectories of animal movement are offered, along with the capability to use automated workflows for processing data from high throughput laboratory experiments.

Users of motion capture software, including researchers who study animal and/or object motion in a laboratory setting, especially for analyses of visual guidance of behavior.

We are not aware of any other R (or otherwise open source) packages that import & wrangle data from Optitrack's `Motive` or the Straw Lab's `Flydra`. Although there are other R packages that handle animal movement trajectory data, these packages tend to be geared towards analyses of large-scale, geospatial movement data (e.g. `trajectories` or various R packages in the `AniMove` software suite). In contrast, `pathviewR` specifically targets tracking of animals (or other subjects) in laboratory settings (e.g. in a flight tunnel) and offers tools to estimate visual perceptions as subjects move in relation to their surroundings.

`pathviewR`, although not specifically geared towards collecting data from human subjects, is in compliance with rOpenSci's policy.
