### 2020-11-16 #409 @maelle
Thanks @vbaliga and co-authors for your submission! I have a few minor comments/questions below, and will now start looking for reviewers.

### Editor checks:

---

#### Editor comments

* In https://vbaliga.github.io/pathviewR/articles/data-import-cleaning.html#data-from-other-sources-1 I would expect to read whether you'd welcome additions of further data import functions, and how an user should request/contribute such a function.

* Why is the CITATION TBD, why not add a citation of the manual for now?

* Why do you use so many continuous integration services, since GitHub Actions would handle all operating systems?

---

Reviewers:
Due date:
