### 2020-11-18 bd86c0b
Create issue template for custom import functions
