+++ .github/ISSUE_TEMPLATE/feature-request--import-function-for-file-type.md
@@ -1,0 +1,23 @@
+---
+name: 'Feature request: import function for file type'
+about: 'Ask us about creating an import function for a specific file type '
+title: "[filetype_request]"
+labels: enhancement
+assignees: ''
+
+---
+
+**Is your feature request related to a specific file type? Please describe.**
+A clear and concise description of what the problem is. What is the file type and are the files created by a specific program or software suite?
+
+**Can you provide an example file of the specified type?**
+Data on public repositories are preferred. If that is not an option, please contact one of the `pathviewR` authors directly.
+
+**Describe the solution you'd like**
+A clear and concise description of what you want to happen when the data are imported from the file.
+
+**Does `as_viewr()` not work?**
+We are very happy to create new functions for additional file types, but we'd also like to know if our current workaround does not work for you. Please provide a description of whether `as_viewr()` can work or if it fails in this use case.
+
+**Additional context**
+Add any other context or screenshots about the feature request here.
