### 2020-11-18 688c198
address editor comments
add citation; add info on how to request custom import functions via github issues; rebuild documentation
