### 2020-11-18 #409 @vbaliga
Hi @maelle,

Thank you for looking over `pathviewR` and providing us with feedback. My co-authors (@scienceisfiction and @epress12) and I worked together earlier today to address each of your comments, and a few changes to our repo have been made in https://github.com/vbaliga/pathviewR/commit/688c198466d1a88a7ff07099263369f6fa39b06c.  Here is our reply to each of your specific comments:

> In https://vbaliga.github.io/pathviewR/articles/data-import-cleaning.html#data-from-other-sources-1 I would expect to read whether you'd welcome additions of further data import functions, and how an user should request/contribute such a function.

Thanks for pointing this out. We very much welcome users to reach out to us regarding further import functions. We created [an Issue template](https://github.com/vbaliga/pathviewR/issues/new/choose) in our repository to allow users to request further import functions, and a link to our Issues templates is now provided in the section of the vignette that you specified.

> Why is the CITATION TBD, why not add a citation of the manual for now?

Thanks for catching this! We have now added a citation of the Manual of this package. It appears both as a `CITATION` file within `/inst` as well as in the Citation section of our readme.

> Why do you use so many continuous integration services, since GitHub Actions would handle all operating systems?

Please forgive our ignorance. We use Travis CI because I had the impression that it was emphasized in the rOpenSci devguide and because of my familiarity with it in authoring a previous rOpenSci package ([ropensci/workloopR](https://github.com/ropensci/workloopR)). We added GitHub Actions because we were interested in seeing what it had to offer. Admittedly, we are less familiar with GitHub Actions' breadth of coverage (moreover, it is a newer service, right?). Would you advise we use only GitHub Actions? Or is it okay that we continue using both Travis CI and GitHub Actions?

> Please add this badge to the package README

The badge has been added. Thanks again for considering our package!

Best regards,
Vikram
🐢
