### 2020-11-18 #409 @maelle
:wave: @vbaliga, @scienceisfiction and @epress12! Thanks for your work!

* Regarding the issue template, awesome! Would you also welcome people to _contribute_ functions and if so what would be the process? Maybe that could be a question in the template?

* Regarding continuous integration yes that's hard to keep up. In https://devguide.ropensci.org/ci.html#whichci we've now put GitHub Actions first which is admittedly not a very obvious way to recommend it. We recommend it over Travis now because it has all operating systems, and because Travis is changing its pricing. E.g. https://www.jeffgeerling.com/blog/2020/travis-cis-new-pricing-plan-threw-wrench-my-open-source-works And since GitHub Actions has support for Windows, you can do without Appveyor. Hard to tell what we'll all be using in a few years :sweat_smile:

I've found a reviewer that I'll assign soon, and am still looking for the second reviewer.
