### 2020-11-18 7c4c32d
try expanding github actions coverage
also disable travis ci
