+++ .github/ISSUE_TEMPLATE/feature-request--import-function-for-file-type.md
@@ -19,5 +19,8 @@ A clear and concise description of what you want to happen when the data are imp
 **Does `as_viewr()` not work?**
 We are very happy to create new functions for additional file types, but we'd also like to know if our current workaround does not work for you. Please provide a description of whether `as_viewr()` can work or if it fails in this use case.
 
+**Are you interested in contributing a function and/or code yourself?**
+If so, please feel free to make a pull request for us to review.
+
 **Additional context**
 Add any other context or screenshots about the feature request here.
