+++ .travis.yml
@@ -1,35 +1,0 @@
-# R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
-
-language: R
-sudo: false
-r:
-  - oldrel # see if this works
-  - release
-  - devel
-cache: packages
-
-r_packages:
-  - R.matlab
-  - data.table
-  - tidyverse
-  - magrittr
-  - dplyr
-  - stringr
-  - tibble
-  - tidyr
-  - fANCOVA
-  - purrr
-  - ggplot2
-  - tidyselect
-  - cowplot
-  - knitr
-  - rmarkdown
-
-r_check_args: "--no-build-vignettes --no-manual --as-cran"
-
-warnings_are_errors: true
-# change the above to true if there are no warnings
-# or set to false if minor warnings should be ignored (temporarily)
-
-after_success:
-  - Rscript -e 'covr::codecov()'
