+++ .github/workflows/workflow.yml
@@ -1,0 +1,11 @@
+steps:
+- uses: actions/checkout@master
+- uses: codecov/codecov-action@v1
+  with:
+    token: ${{ secrets.CODECOV_TOKEN }} # not required for public repos
+    file: ./coverage.xml # optional
+    files: ./coverage1.xml,./coverage2.xml # optional
+    flags: unittests # optional
+    name: codecov-umbrella # optional
+    fail_ci_if_error: true # optional (default = false)
+    verbose: true # optional (default = false)
