### 2020-11-18 d358c9d
disable optional steps in workflow.yml
