### 2020-11-18 7342a4d
revise triggering of codecov with github actions
