+++ .github/workflows/workflow.yml
@@ -1,3 +1,0 @@
-steps:
-- uses: actions/checkout@master
-- uses: codecov/codecov-action@v1
