### 2020-11-18 ae5bf4e
make standalone codecov yaml
