### 2020-11-19 f3a787b
remove travis badge and redocument
