### 2020-11-19 #409 @vbaliga
Hi @maelle,

Thanks very much for your advice. We've expanded our issue template to welcome people to contribute functions and added a question to the template accordingly (see https://github.com/vbaliga/pathviewR/commit/7c4c32d977baf966da76366c88e746630cf17f43). We have also now shifted to using GitHub Actions exclusively for CI using the "standard" check across Mac/Win/Linux and revised our method of updating Codecov through Actions as well (see https://github.com/vbaliga/pathviewR/commit/f3a787b4350c12e8ec707e3e0a4cef9b6a7dca6f).

Thanks again!
Vikram
🐢
