+++ .github/ISSUE_TEMPLATE/general-feature-request.md
@@ -1,0 +1,25 @@
+---
+name: 'Bug fix request'
+about: 'Notify us of a bug so we can fix it'
+title: "[bug_fix_request]"
+labels: enhancement
+assignees: ''
+
+---
+
+**Which function(s) is failing?**
+A clear and concise description of what the problem is.
+
+**Can you provide a reproducible example of the failure?**
+Data on public repositories are preferred with steps to reproduce the failure. If that is not an option, please contact one of the `pathviewR` authors directly.
+
+**What error message(s) are you receiving, if any?**
+
+**Expected behavior or describe the solution you'd like**
+A clear and concise description of what you expected to happen or what you would like to happen.
+
+**What file type are you using?**
+.csv from Motive, .mat from Flydra or other source using `as_viewr()` to import. If other, please describe.
+
+**Additional context**
+Add any other context or screenshots about the feature request here.
