+++ .github/ISSUE_TEMPLATE/general-feature-request.md
@@ -1,25 +1,23 @@
 ---
-name: 'Bug fix request'
-about: 'Notify us of a bug so we can fix it'
-title: "[bug_fix_request]"
+name: 'General Feature Request'
+about: 'Let us know if you have an idea to improve `pathviewR`'
+title: "[genera_feature_request]"
 labels: enhancement
 assignees: ''
 
 ---
 
-**Which function(s) is failing?**
-A clear and concise description of what the problem is.
+**Please describe the nature of your feature request.**
+Is the request to improve the tool(s) currently within `pathviewR` or to provide a tool that adds to the package's capabilities?
 
-**Can you provide a reproducible example of the failure?**
-Data on public repositories are preferred with steps to reproduce the failure. If that is not an option, please contact one of the `pathviewR` authors directly.
+**Please explain the feature you would like to use in `pathviewR`.**
+Briefly describe the goal of the feature and/or which current aspects of `pathviewR` relate to your request.
 
-**What error message(s) are you receiving, if any?**
+**Can you provide an example file to help explain your request?**
+Data on public repositories are preferred. If that is not an option, please contact one of the `pathviewR` authors directly.
 
-**Expected behavior or describe the solution you'd like**
-A clear and concise description of what you expected to happen or what you would like to happen.
+**Are you interested in contributing a function and/or code yourself?**
+If so, please feel free to make a pull request for us to review.
 
-**What file type are you using?**
-.csv from Motive, .mat from Flydra or other source using `as_viewr()` to import. If other, please describe.
-
-**Additional context**
+**Additional context** 
 Add any other context or screenshots about the feature request here.
