### 2020-11-30 #409 @maelle
Thanks a lot @marcosci for agreeing to review! Your review is due on 2020-12-21.

Please find links to our [reviewer guide](https://devguide.ropensci.org/reviewerguide.html) and [review template](https://devguide.ropensci.org/reviewtemplate.html).
