### 2020-12-04 #409 @stefaniebutland
Hi @asbonnetlebrun. I'm rOpenSci's Community Manager. If you would like to join our Slack group, please email me at `stefanie@ropensci.org`; I was not able to find your email address.
