### 2020-12-10 #409 @asbonnetlebrun
I don’t have any working relationship with any of the package authors.

There are examples for almost all exported functions in R help. Exceptions are the `calc_sf_box()`, `calc_sf_V()`, `calc_vis_angle_box()` and `calc_vis_angle_V()` functions (and standardised_tunnel, although there is a reason given in the R help – because no example data is provided with pathviewR to run the code on).

Only in the DESCRIPTION file, but not in the README.

See details in my comments about some potential coding mistakes.

All tests pass on my machine as they have been designed (but see in my comments for one test that likely should be re-written).

---

I want to first point out that my background is on animal movement data, but not with motion capture cameras (I focus on tracking of wild animals, using GPS or similar loggers), so I was not fully familiar with the interests/culture of the field.

I thought the package was well documented overall, and the three vignettes in particular really helped me understand what the package was for. That being said, I think it would still benefit from some additional information in the different documents (vignettes, README, R help). See my comments below from some specific points that would be good to add.

I also very much appreciated the effort made to make it all work with pipes.

In the rest of my review, I’ll first focus on comments on the three vignettes to help clarify some points to the users. Then I’ll focus on the code of the visual perception functions, where I think I found some mistakes and that would be worth checking.

**Data import and cleaning**
My understanding is that the “Relabeling axes” and “gathering data columns” parts are only relevant in the case of Motive data (Flydra data already comes in the correct format, and `as_viewr()`, the function to import other types of data also takes data in the correct format and already relabels the columns). Maybe this could be made a bit clearer in the vignette?

Maybe that was me not being very familiar with the kind of experiments the package is relevant for, but I had trouble to understand when we would use the `standardize_tunnel()` function or the `rotate_tunnel()` function. Maybe you could provide some examples when these functions to be applied? Also I struggled to understand the `standardize_tunnel()` function, maybe particularly because there was no example in the help file. When would the landmarks already be in the data? I guess it will never be the case with Flydra data, considering that the `read_flydra_mat()` function allows to input only a single `subject_name.` Is that why it doesn’t need to be rotated, or is that something arising from the Flydra software?

Also, considering how the `select_x_percent()` function works (by selecting a certain percentage of the tunnel length on each site of (0,0,0) along the length axis), shouldn’t it be more appropriate to say that the (0,0,0) must be at the centre of the region of interest, rather than at the centre of the tunnel?

Minor point: the link to the vignette for managing frame gaps is missing in the text.

**Managing frame gaps**
Very minor point: could it be clearer in the `visualize_frame_gap_choice()` help or in the vignette that the loops argument represents the maximum frame gap considered (i.e. that each loop represents an increment of 1 on the frame gap value)?

Could there be cases when frame gaps can vary between devices (i.e. if I understood well in the case of the Motive data, between subjects)? If so, would it be relevant to allow the autodetect approach to be applied to each subject separately? (or maybe that is not relevant...)

**Visual perception**
This is where I struggled, but maybe that is because I don’t come from the specific field of application of this package. I felt like some more details could be included in the vignette. Maybe start with a few examples of experiments where this package can be useful. Maybe you could say right at the start that these functions are relevant for experiments with animals flying in tunnels with visual stimuli consisting of sine wave gratings on the tunnel walls?

Also, my understanding is that here you implement only two cases:
-	Motive example: birds flying through a V-shaped tunnel, with visual stimuli on each side of the tunnel
-	Flydra example: birds flying through a rectangular tunnel, with visual stimuli on the side and front walls
So maybe make it clear that this section only applies to these two specific cases?

On this, a question: is there any plan to code more situations or are these the typical situations for these kinds of experiment? If you welcome suggestions from users of other settings, maybe you could say that in the vignette in the same way you mention that you are happy to work towards the inclusion of more data types?

More generally, I would have appreciated some definition of what you call “spatial frequency” and “visual angle”. Although I understand that these might be obvious for people in the field (which are ultimately the target users of the package), but I feel like these terms (visual angle, and in particular spatial frequency) are of such broad usage that they might deserve some better description of what they mean in the context of the package. Maybe include a diagram in the vignette for users to visualise what these values represent?

In particular, the Value section in the R help for the `calc_sf_box()` function seems to mention that the spatial frequency is the number of cycles per degree of visual angle. Maybe this info could be included in the vignette (and in the Description section of the function help)?

**Comments on the code of the visual perception functions**
I seem to understand that the functions to calculate visual angle and spatial frequency in the box example don’t handle the front wall (only the side walls), am I right? Is there any reason for this? If the front wall is not relevant, maybe there is no need to include the option to add parameters for it in the `insert_treatments()` function?

My understanding of the visual angle is as follows: let’s consider an isosceles triangle with one vertex at the bird’s eye, with the side of unique length being on the wall and of a length equal to the wavelength of the stimulus, and the third vertex to be at the bird’s eye, with the bisector of the angle at the bird’s eye being the segment of shortest distance between the bird and the wall (would have been easier with a diagram, sorry). Then the angle at the bird’s eye should be the visual angle – right?

If I am correct, let’s dig into the `calc_sf_box()`, `calc_sf_V()`, `calc_vis_angle_box()` and `calc_vis_angle_V()` functions.
My understanding in all these functions is that `min_dist_pos` is the distance between the bird and the positive wall, and `min_dist_neg`, between the bird and the negative wall.
Calculations for these values are redundant between `calc_sf_box()` and `calc_vis_angle_box()`, as well as between `calc_sf_V()` and `calc_vis_angle_V()`. I would recommend to write functions to calculate these distances, that would then be called in the `calc_*` functions to ensure consistency. In particular, currently the `min_dist_neg` calculations differ between `calc_sf_box()` and `calc_vis_angle_box()`, the correct one being in `calc_sf_box()`. The effect of the error in `calc_vis_angle_box()` is that once the widths become positive, the distance to the negative wall start to decrease the further away we get from this wall. This can be visualised by slightly tweaking the final bit of code in the “Visual perception” vignette:

```
ggplot(flydra_box_angle, aes(x = position_width, y = position_length)) +
  geom_point(aes(colour = min_dist_neg), size = 2) +
  scale_colour_viridis_c() +
  coord_fixed() +
  geom_segment(aes(x = -0.5,        # negative wall
                   y = -0.5,
                   xend = -0.5,
                   yend = 0.5)) +
  geom_segment(aes(x = 0.5,         # positive wall
                   y = -0.5,
                   xend = 0.5,
                   yend = 0.5))
```

On that note, there is no need for an `ifelse()` to calculate these distances, these lines could simply replace L9-14 in calc_sf_box() and `calc_vis_angle_box()`:

```
  obj_name$min_dist_pos <- obj_name$pos_wall - obj_name$position_width
  obj_name$min_dist_neg <- abs(obj_name$neg_wall + obj_name$position_width)
```
Can you please correct the calculations, and adapt the test-calc_vis_angle_box.R file?

And also on this, I noticed that the user can supply negative values for the `neg_wall` and `pos_wall` arguments in the `insert_treatments()` function, which in this case would lead to spurious distances being calculated. Maybe insert a warning/error message if that is the case, and add the appropriate tests?

Finally, still on distances, inspecting the resulting distances to the negative and positive walls from a range of positive and negative width positions seem to indicate that the distance calculations for the other functions are correct.

**Spatial frequency**
Another point, on the spatial frequency. If I understand it well (being the number of cycles per degree of visual angle), I think there is an error in L34-35 of `calc_sf_V()` and L15-16 of `calc_sf_box()`.

It currently is:
```
deg_dist_pos <- 2 * obj_name$min_dist_pos * tan(deg_2_rad(1))
deg_dist_neg <- 2 * obj_name$min_dist_neg * tan(deg_2_rad(1))
```
I think, unless I got the trigonometry wrong, that this represents the number of cycles per 2 degrees of visual angle, and therefore should be changed to:
```
deg_dist_pos <- 2 * obj_name$min_dist_pos * tan(deg_2_rad(1/2))
deg_dist_neg <- 2 * obj_name$min_dist_neg * tan(deg_2_rad(1/2))
```

Just one final broader question, don’t these calculations assume that the bird is parallel to the length axis of the tunnel? Is that not important/common practice not to use the rotation information?

Many thanks for your contribution, and I hope this will helps!
