### 2020-12-14 #409 @maelle
Thanks a ton @marcosci for your review! :smile_cat:

> I tried the package out with a .csv file from my phone that I tracked while going for a walk and everything worked just fine.

Such a neat idea!

Regarding the package name, yes we _recommend_ all lower-case, especially if the package isn't on CRAN yet. I myself renamed my Ropenaq package after review and I don't regret doing it.
