### 2020-12-27 30d83ef
rough work to address visual perception function reviews - variable name changes - function additions: calc_min_dist_v, calc_min_dist_box, get_vis_angle, get_sf, get_tf, get_pattern velocity, get_image_expansion
