### 2020-12-28 7f666ce
all in one visual perception functioin
