### 2020-12-28 c4c4ac6
add tunnel dimensions to arguments of insert_treatments() - add min_dist_end to calc_min_dist functions using these new arguments
