+++ example_scripts/vis_percept_reviews.R
@@ -1,7 +1,94 @@
 ## work space for addressing visual perception functions reivews
-## Things to address:
+
+## change argument for vertex_angle to mean the actual angle of the V in the
+## V-shaped tunnel. Build into insert_treatments() a calculation that makes it
+## work with calc_min_dist functions
+
+## insert_treatments()
+## use tunnel_length argument so that front/back wall distances can be calculated
+
+insert_treatments <- function(obj_name,
+                              perch_2_vertex = NULL,
+                              vertex_angle = NULL, # actual angle of vertex
+                              tunnel_width = NULL,
+                              tunnel_length = NULL,
+                              stim_param_lat_pos = NULL,
+                              stim_param_lat_neg = NULL,
+                              stim_param_end_pos = NULL,
+                              stim_param_end_neg = NULL,
+                              treatment = NULL){
+
+  ## Check that it's a viewr object
+  if (!any(attr(obj_name,"pathviewR_steps") == "viewr")){
+    stop("This doesn't seem to be a viewr object")
+  }
+
+  ## Check that get_full_trajectories has been run prior to use
+  if (!any(attr(obj_name, "pathviewR_steps") == "full_trajectories")){
+    stop("Run get_full_trajectories() prior to use")
+  }
+
+
+  ## Translate arguments into variables at beginning of data frame
+  ##
+  ## figure out how to do it such that only the arguments supplied are added to
+  ## the data frame
+  if (attr(obj_name, "import_method") == "motive"){
+    obj_name <- tibble::add_column(obj_name, .before = "frame",
+                                   perch_2_vertex = perch_2_vertex,
+                                   vertex_angle = deg_2_rad(vertex_angle/2),
+                                   tunnel_width = tunnel_width,
+                                   tunnel_length = tunnel_length,
+                                   stim_param_lat_pos = stim_param_lat_pos,
+                                   stim_param_lat_neg = stim_param_lat_neg,
+                                   stim_param_end_pos = stim_param_end_pos,
+                                   stim_param_end_neg = stim_param_end_neg,
+                                   treatment = treatment)
+  } else if (attr(obj_name, "import_method") == "flydra"){
+    obj_name <- tibble::add_column(obj_name, .before = "frame",
+                                   tunnel_width = tunel_width,
+                                   tunnel_length= tunnel_length,
+                                   stim_param_lat_pos = stim_param_lat_pos,
+                                   stim_param_lat_neg = stim_param_lat_neg,
+                                   stim_param_end_pos = stim_param_end_pos,
+                                   stim_param_end_neg = stim_param_end_neg,
+                                   treatment = treatment)
+  }
+
+  ## Add arguments into metadata......surewhynot
+  if (attr(obj_name, "import_method") == "motive"){
+    attr(obj_name, "perch_2_vertex") <- perch_2_vertex
+    attr(obj_name, "vertex_angle") <- vertex_angle
+    attr(obj_name, "tunnel_width") <- tunnel_width
+    attr(obj_name, "tunnel_length") <- tunnel_length
+    attr(obj_name, "stim_param_lat_pos") <- stim_param_lat_pos
+    attr(obj_name, "stim_param_lat_neg") <- stim_param_lat_neg
+    attr(obj_name, "stim_param_end_pos") <- stim_param_end_pos
[…]
