+++ example_scripts/ropensci_reviews_checklist.md
@@ -1,0 +1,291 @@
+Checklist for rOpenSci reviews - round 1
+================
+
+
+This Markdown file will contain items from each of our reviews along with a
+checklist to ensure that we address each item. Reviews may not be reproduced in
+full, but all of the items that require action from us should be included here
+(please add to this document if anything is missing).
+
+Ultimately, the contents of this file will be transformed into our Response to
+Reviews.
+
+As we address items, we should:  
+- check it off in this file  
+- add in the name of the commit to this file using the first 7
+characters of the commit ID
+
+For example:  
+- [x] consolidate visual perception functions into one function
+[(7f666ce)](https://github.com/vbaliga/pathviewR/commit/7f666ce6e9eee09545ddd9e30d18c0c87ab7eeaf)  
+- [ ] get a million dollars
+
+The first item was addressed and the corresponding commit has been
+linked. Note that the item has been checked off via adding an “x” to “[ ]”.
+
+The second item is...a work in progress...
+
+## Review from Reviewer 1 (asbonnetlebrun)
+
+### Documentation
+#### **Examples** for all exported functions in R Help that run successfully locally
+
+> There are examples for almost all exported functions in R help. Exceptions are
+the calc_sf_box(), calc_sf_V(), calc_vis_angle_box() and calc_vis_angle_V()
+functions (and standardised_tunnel, although there is a reason given in the R
+help – because no example data is provided with pathviewR to run the code on).
+
+Items for us:  
+- [ ] consider and/or make examples for calc_sf_box(), calc_sf_V(),
+calc_vis_angle_box() and calc_vis_angle_V() (OR, make examples for whatever
+visual perception function(s) we make)
+
+#### **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with URL, BugReports and Maintainer (which may be autogenerated via Authors@R).
+
+> Only in the DESCRIPTION file, but not in the README. 
+
+Items for us:  
+- [ ] add contribution guidelines (links to Issues pages) in README
+
+#### **Functionality**
+*We will address these items in the `Review Comments` section below*
+
+> Should the author(s) deem it appropriate, I agree to be acknowledged as a
+package reviewer ("rev" role) in the package DESCRIPTION file.
+
+Items for us:  
+- [ ] add asbonnetlebrun as a "rev" in DESCRIPTION
+
+### Review Comments
+
+#### Data import and cleaning 
+ 
+> My understanding is that the “Relabeling axes” and “gathering data columns”
+parts are only relevant in the case of Motive data (Flydra data already comes in
+the correct format, and as_viewr(), the function to import other types of data
+also takes data in the correct format and already relabels the columns). Maybe
+this could be made a bit clearer in the vignette?
+
+Items for us:   
+- [ ] clarify language of this vignette to indicate that relabeling & gathering
+are only necessary in certain cases (e.g. using Motive data)
+
+> Maybe that was me not being very familiar with the kind of experiments the
+package is relevant for, but I had trouble to understand when we would use the
+standardize_tunnel() function or the rotate_tunnel() function. Maybe you could
+provide some examples when these functions to be applied? Also I struggled to
+understand the standardize_tunnel() function, maybe particularly because there
+was no example in the help file. When would the landmarks already be in the
+data? I guess it will never be the case with Flydra data, considering that the
+read_flydra_mat() function allows to input only a single subject_name. Is that
+why it doesn’t need to be rotated, or is that something arising from the Flydra
+software?
+
+Items for us:   
+- [ ] in this vignette: clarify the circumstances under which standardization is
+needed and what types of landmarks are appropriate (consider adding a figure?)  
+- [ ] in the Help file for `standardize_tunnel()`: clarify this function's use 
+cases and perhaps link to the vignette itself?  
[…]
