### 2020-12-29 5252918
add contributing guides
via rOpenSci templates
