+++ vignettes/data-import-cleaning.Rmd
@@ -18,6 +18,20 @@ knitr::opts_chunk$set(
 )
 ```
 
+## Overview
+
+Raw movement data, including those from motion capture systems, may have a
+variety of issues. These raw data often contain noise or artifacts from the
+recording session, which may not be easily removed via the recording software
+itself. Data may not be organized as “tidy” key-value pairs, the axes and
+overall orientation of the environment may not conform to a standard, and
+individual movement trajectories may be ill-defined. 
+
+`pathviewR` provides functions in R to deal with such problems. This vignette
+will cover the basics of how to import raw data and how to "clean" data to
+prepare it for statistical analyses.
+
+
 ## Data import via `pathviewR`
 
 Data can be imported via one of three ways:  
@@ -89,7 +103,7 @@ attr(motive_data, "frame_rate")
 
 `.mat` files exported from Flydra can be imported via `read_flydra_mat()`.
 
-Note that a subject name is required. Only one name can be added and it will
+Note that a `subject_name` is required. Only one name can be added and it will
 be used throughout the resultant `tibble`.
 
 ```{r import_flydra}
