### 2020-12-29 f52b8d8
Add additional calculations to perception functions - visual angles to end walls - additional temporal frequency calculations for different types of animal motion. - at this point, a more advanced analysis can be performed on data captured in a rectangular tunnel due to simpler math.
