### 2020-12-29 266ff53
Add temporal frequency and pattern velocity for planar movements in box-shaped tunnel
