### 2020-12-30 591dd50
update data import vignette
addressing additonal items from our ropensci reviews
