+++ vignettes/data-import-cleaning.Rmd
@@ -27,20 +27,38 @@ itself. Data may not be organized as “tidy” key-value pairs, the axes and
 overall orientation of the environment may not conform to a standard, and
 individual movement trajectories may be ill-defined. 
 
-`pathviewR` provides functions in R to deal with such problems. This vignette
-will cover the basics of how to import raw data and how to "clean" data to
-prepare it for statistical analyses.
+`pathviewR` provides functions in R to deal with such problems (i.e. "clean"
+them). This vignette will cover the basics of how to import raw data and how to
+clean data to prepare it for visualization and/or statistical analyses.
 
+
+## What do movement data sets look like?
+
+At minimum, movement data provide information on a subject or object's position
+over time. These data are typically supplied in three dimensions (e.g. x, y, z),
+with position in each dimension sampled at a particular rate (e.g. 100 Hz).
+Different recording software may provide additional features, such as the
+ability to track multiple subjects simultaneously, information on subjects'
+rotation, tracking of "rigid body" elements, or even the ability to apply Kalman
+filters.
+
+A central goal of `pathviewR` is to take data from different sources (so far:
+Motive and Flydra), re-organize them into a common format that can be wrangled
+in R, clean them up a bit, and get them ready for visualization and/or
+statistical analyses. We'll first cover what's included in Motive and in Flydra
+data and how `pathviewR` handles these. Should you have data from another
+source, our `as_viewr()` function will allow you to bring it into the
+`pathviewR` framework.
 
 ## Data import via `pathviewR`
 
-Data can be imported via one of three ways:  
+Data can be imported via one of three functions:  
 
-- `read_motive_csv()` imports data from `.csv` files exported by 
-[Optitrack's Motive](https://optitrack.com/software/motive/) software  
+- `read_motive_csv()` imports data from `.csv` files that have been exported
+from [Optitrack's Motive](https://optitrack.com/software/motive/) software
 
-- `read_flydra_mat()` imports data from  `.mat` files exported from 
-[Flydra](https://github.com/strawlab/flydra)  
+- `read_flydra_mat()` imports data from  `.mat` files that have been exported
+from [Flydra](https://github.com/strawlab/flydra)
 
 - `as_viewr()` can be used to handle data from other sources  
 
@@ -82,6 +100,15 @@ motive_data <-
 motive_data
 ``` 
 
+A key thing to note is that these data, as stored in Motive CSVs, are not
+"tidy". Each frame occupies one row, but what that also means is that the
+rotation and position values for the various subjects take up 24 columns! This
+format not only makes plotting data more difficult in base R, `ggplot2`, and
+`rgl`, but also makes other aspects of data wrangling more difficult. In a later
+step, we will 'gather' these data into key-value pairs so that e.g. all
+length-wise position values are in one column, all width-wise are in
+another...etc.
+
 Metadata are stored as attributes. We won't go through all of these, but here
 are a couple important ones.
 
@@ -98,6 +125,11 @@ attr(motive_data, "data_types_simple")
 ## Frame rate
 attr(motive_data, "frame_rate")
 ```
+
+Storing such metatdata in the attributes is a key feature of `pathviewR`. These
+metadata may not be as immediately as important as the time series of position
+or rotation, but they can provide important experimental information such as the
+date & time of capture and the units of the position data (here, meters).
 
 ### Flydra Matlab files
 
@@ -120,7 +152,13 @@ flydra_data <-
 ## Similarly, this produces a tibble with important 
 ## metadata as attributes
 flydra_data
+
+attr(flydra_data, "frame_rate")
 ```
+
+Note that unlike the example Motive data, the Flydra data are already organized
[…]
