### 2020-12-30 bcf6424
add details about loops to vignette and help file of manage frame gaps
