### 2020-12-30 62b63b7
small language change for autodetect in max frame gap vignette
