### 2020-12-30 984ffd8
add in some commits to checklist
i neglected to do this yesterday
