### 2020-12-30 b29dd4b
test out and adjust functions with Flydra data set - get_tf() still needs some work on "if" section
