+++ example_scripts/vis_percept_reviews.R
@@ -46,7 +46,7 @@ insert_treatments <- function(obj_name,
                                    treatment = treatment)
   } else if (attr(obj_name, "import_method") == "flydra"){
     obj_name <- tibble::add_column(obj_name, .before = "frame",
-                                   tunnel_width = tunel_width,
+                                   tunnel_width = tunnel_width,
                                    tunnel_length= tunnel_length,
                                    stim_param_lat_pos = stim_param_lat_pos,
                                    stim_param_lat_neg = stim_param_lat_neg,
@@ -175,11 +175,10 @@ calc_min_dist_v <- function(obj_name,
            # return original min_dist_neg calculation
 
   ## For minimum distances to end walls (assuming animal locomotes forward)
-  if (obj_name$direction == 1){
-    obj_name$min_dist_end <- (tunnel_length/2 - position_length)
-  } else if (obj_name$direction == -1) {
-    obj_name$min_dist_end <- abs(-tunnel_length/2 - position_length)
-  }
+  obj_name$min_dist_end <-
+    ifelse(obj_name$end_length_sign == 1,
+           obj_name$tunnel_length/2 - obj_name$position_length,
+           obj_name$tunnel_length/2 + obj_name$position_length)
 
 
   ## Leave note that minimum distances were calculated
@@ -200,8 +199,6 @@ calc_min_dist_v <- function(obj_name,
 
 
 ## min_dist function for box shaped chamber
-## I believe the ifelse statements for calculating min_dist are necessary - the
-## reviewer's suggestion doesn't work as required.
 
 calc_min_dist_box <- function(obj_name){
 
@@ -217,24 +214,14 @@ calc_min_dist_box <- function(obj_name){
 
   ## Calculate minimum distance to each wall from positive or negative sides of
   ## tunnel
-  obj_name$min_dist_pos <-
-    ifelse(obj_name$position_width >= 0, # if in positive side of tunnel
-           obj_name$pos_wall - obj_name$position_width, # TRUE
-           obj_name$pos_wall + abs(obj_name$position_width) # FALSE
-    )
-
-  obj_name$min_dist_neg <-
-    ifelse(obj_name$position_width <= 0, # if in negative side of tunnel
-           abs(obj_name$neg_wall + obj_name$position_width), # TRUE
-           abs(obj_name$neg_wall) + obj_name$position_width # FALSE
-    )
+  obj_name$min_dist_pos <- obj_name$tunnel_width/2 - obj_name$position_width
+  obj_name$min_dist_neg <- obj_name$tunnel_width/2 + obj_name$position_width
 
   ## For minimum distances to end walls (assuming animal locomotes forward)
-  if (obj_name$direction == 1){
-    obj_name$min_dist_end <- (tunnel_length/2 - position_length)
-  } else if (obj_name$direction == -1) {
-    obj_name$min_dist_end <- abs(-tunnel_length/2 - position_length)
-  }
+  obj_name$min_dist_end <-
+    ifelse(obj_name$end_length_sign == 1,
+           obj_name$tunnel_length/2 - obj_name$position_length,
+           obj_name$tunnel_length/2 + obj_name$position_length)
 
   ## Leave note that minimum distances were calculated
   attr(obj_name, "pathviewR_steps") <- c(attr(obj_name, "pathviewR_steps"),
[…]
