### 2020-12-31 4100420
I should commit more frequently... - fixed some typos - commented out faulty checks in get_vis_angle() and get_pattern_vel() - new function get_pattern_velocity_direction() - tested all functions on flydra dataset.. all seem to be working properly
