+++ example_scripts/vis_percept_reviews.R
@@ -242,8 +242,8 @@ get_vis_angle <- function(obj_name){
   }
 
   ## Check that calc_min_dist() has been run
-  #if (!any(attr(obj_name,"pathviewR_steps") == "min_dist_v_calculated" |
-   #                                            "min_dist_box_calculated")){
+  #if (!any(attr(obj_name,"pathviewR_steps") == "min_dist_v_calculated" ||
+   #   !any(attr(obj_name,"pathviewR_steps") == "min_dist_box_calculated"))){
     #stop("Please calculate minimum distances prior to use")
   #}
 
@@ -323,7 +323,7 @@ get_tf <- function(obj_name,
 
   ## Temporal frequency (cycles/second)instantaneous velocity of the subject and
   ## the arc length of the cycle from the subject's perspective
-  if (any(attr(obj_name, "pathviewR_steps")) == "min_dist_box_calculated"){
+  if (any(attr(obj_name, "pathviewR_steps") == "min_dist_box_calculated")){
 
         ## Lateral wall 1D (axial) temporal frequencies
     # tf of positive wall from forward movement
@@ -367,7 +367,7 @@ get_tf <- function(obj_name,
     obj_name$tf_planar_end <- abs(obj_name$end_planar_inst_vel)/
       (obj_name$min_dist_end * obj_name$vis_angle_end_rad)
 
-  } else if (any(attr(obj_name, "pathviewR_steps")) == "min_dist_v_calculated"){
+  } else if (any(attr(obj_name, "pathviewR_steps") == "min_dist_v_calculated")){
         ## Lateral wall 1D (axial) temporal frequencies
     # tf of positive wall from forward movement
     obj_name$tf_forward_pos <- abs(obj_name$length_inst_vel)/
@@ -380,11 +380,9 @@ get_tf <- function(obj_name,
   ### analysis of how multi-dimensional motion affects temporal frequency is in
   ### development for V-shaped tunnel configurations
 
-  ## Leave a note that visual angles were calculated
+  ## Leave a note that temporal frequencies  were calculated
   attr(obj_name, "pathviewR_steps") <- c(attr(obj_name, "pathviewR_steps"),
                                          "tf_calculated")
-
-
 
   return(obj_name)
 }
@@ -405,47 +403,48 @@ get_pattern_vel <- function(obj_name){
   }
 
   ## Check that get_sf() and get_tf() have been run
-  if (!any(attr(obj_name, "pathviewR_steps") == "sf_calculated" |
-                                                "tf_calculated")){
-    stop("Please run get_sf() and get_tf() prior to use")
-  }
+  #if (!any(attr(obj_name, "pathviewR_steps") == "sf_calculated" ||
+   #   !any(attr(obj_name, "pathviewR_steps") == "tf_calculated"))){
+    #stop("Please run get_sf() and get_tf() prior to use")
+  #}
 
-  if (any(attr(obj_name, "pathviewR_steps")) == "min_dist_box_calculated"){
-      ## Lateral wall pattern velocities ##
+  ## For box-shaped tunnel configuration
+  if (any(attr(obj_name, "pathviewR_steps") == "min_dist_box_calculated")){
+
+    ## Lateral wall pattern velocities ##
   # forward pattern velocity from positive wall
-  obj_name$forward_pattern_vel_pos <- tf_forward_pos/sf_pos
+  obj_name$forward_pattern_vel_pos <- obj_name$tf_forward_pos/obj_name$sf_pos
   # forward pattern velocity from negative wall
-  obj_name$forward_pattern_vel_neg <- tf_forward_neg/sf_neg
[…]
