### 2021-01-02 755df32
change metadata output of calc_min_dist functions and implement new tunnel_config variable argument in subsequent function conditionals
