+++ example_scripts/vis_percept_reviews.R
@@ -59,6 +59,7 @@ insert_treatments <- function(obj_name,
 
   ## Add arguments into metadata......surewhynot
   if (tunnel_config == "v"){
+    attr(obj_name, "tunnel_config") <- tunnel_config
     attr(obj_name, "perch_2_vertex") <- perch_2_vertex
     attr(obj_name, "vertex_angle") <- vertex_angle
     attr(obj_name, "tunnel_width") <- tunnel_width
@@ -69,6 +70,7 @@ insert_treatments <- function(obj_name,
     attr(obj_name, "stim_param_end_neg") <- stim_param_end_neg
     attr(obj_name, "treatment") <- treatment
   } else if (tunnel_config == "box"){
+    attr(obj_name, "tunnel_config") <- tunnel_config
     attr(obj_name, "tunnel_width") <- tunnel_width
     attr(obj_name, "tunnel_length") <- tunnel_length
     attr(obj_name, "stim_param_lat_pos") <- stim_param_lat_pos
@@ -118,7 +120,6 @@ calc_min_dist_v <- function(obj_name,
     obj_name$vertical_2_vertex -
     (abs(obj_name$position_width) / tan(obj_name$vertex_angle))
 
-
   ## Introduce variables for horizontal_2_screen on positive and negative sides
   ## of the tunnel.
   ## horizontal_2_screen refers to the horizontal distance between the bird and
@@ -146,7 +147,6 @@ calc_min_dist_v <- function(obj_name,
     obj_name$horizontal_2_screen_neg * sin(pi/2 - obj_name$vertex_angle)
   # min_dist to negative screen
 
-
   ## When the subject is outside the boundaries created by orthogonal planes to
   ## each wall, erroneous visual angles are calculated.
   ## Therefore we must adjust min_dist values according to position_width
@@ -156,7 +156,6 @@ calc_min_dist_v <- function(obj_name,
     obj_name$vertical_2_vertex * tan(pi/2 - obj_name$vertex_angle)
   obj_name$bound_neg <-
     obj_name$vertical_2_vertex * -tan(pi/2 - obj_name$vertex_angle)
-
 
   obj_name$min_dist_pos <- # overwrite min_dist_pos
     ifelse(obj_name$position_width <= 0 &
@@ -183,18 +182,19 @@ calc_min_dist_v <- function(obj_name,
            obj_name$tunnel_length/2 + obj_name$position_length)
 
 
-  ## Leave note that minimum distances were calculated
-  attr(obj_name, "pathviewR_steps") <- c(attr(obj_name, "pathviewR_steps"),
-                                         "min_dist_calculated")
-
   ## for simplify = TRUE
   obj_simplify$min_dist_pos <- obj_name$min_dist_pos
   obj_simplify$min_dist_neg <- obj_name$min_dist_neg
   obj_simplify$min_dist_end <- obj_name$min_dist_end
 
+  ## return object and add note that minimum distaces were calculated
   if(simplify == TRUE){
+    attr(obj_simplify, "pathviewR_steps") <-
+      c(attr(obj_name, "pathviewR_steps"), "min_dist_calculated")
     return(obj_simplify)
   } else {
+    attr(obj_name, "pathviewR_steps") <-
+      c(attr(obj_name, "pathviewR_steps"), "min_dist_calculated")
     return(obj_name)
   }
 }
@@ -325,58 +325,90 @@ get_tf <- function(obj_name,
   ## the arc length of the cycle from the subject's perspective
 
   ## For box-shaped tunnel
-   if (obj_name$tunnel_config == "box"){
+   if (attr(obj_name, "tunnel_config") == "box"){
+
         ## Lateral wall 1D (axial) temporal frequencies
[…]
