+++ R/utility_functions.R
@@ -2605,33 +2605,34 @@ rm_by_trajnum <- function(obj_name,
 #'
 #' @param obj_name The input viewr object; a tibble or data.frame with attribute
 #'   \code{pathviewR_steps} that includes \code{"viewr"}
-#' @param vertex_height If using a V-shaped chamber, this is the height of the
-#'   vertex. This value should be negative and represents the distance between
-#'   the origin (0,0,0) and the vertex.
-#' @param vertex_angle  If using a V-shaped chamber, this is the angle (in
-#'   degrees) subtended by a vertical axis and the sides of the V-shaped
-#'   chamber. Equivalent to half the angle of the "V". \code{vertex_angle}
-#'   defaults to 45.
-#' @param pos_wall If using a box-shaped chamber, this is the distance in meters
-#'   between the origin (0,0,0) and the positive wall.
-#' @param neg_wall If using a box-shaped chamber, this is the distance in meters
-#'   between the origin (0,0,0) and the negative wall.
-#' @param front_wall If using a box-shaped chamber, this is the distance in
-#'   meters between the origin (0,0,0) and the front wall.
+#' @param tunnel_config The configuration of the experimental tunnel.
+#' Currently, pathviewR supports rectangular "box" and V-shaped tunnel
+#' configurations.
+#' @param perch_2_vertex If using a V-shaped tunnel, this is the vertical
+#' distance between the vertex and the height of the perches. If the tunnel does
+#' not have perches, insert the vertical distance between the vertex and the
+#' height of the origin (0,0,0).
+#' @param vertex_angle  If using a V-shaped tunnel, the angle of the vertex (in
+#' degrees) \code{vertex_angle} defaults to 90.
+#' @param tunnel_width If using a box-shaped tunnel, the width of the tunnel.
+#' @param tunnel_length The length of the tunnel.
+#' @param stim_param_lat_pos The size of the stimulus on the lateral positive
+#' wall of the tunnel. Eg. for 10cm wide gratings,
+#' \code{stim_param_lat_pos} = 0.1.
+#' @param stim_param_lat_neg The size of the stimulus on the lateral negative
+#' wall of the tunnel..
+#' @param stim_param_end_pos The size of the stimulus on the end positive
+#' wall of the tunnel.
+#' @param stim_param_end_neg The size of the stimulus on the end negative
+#' wall of the tunnel.
 #' @param treatment The name of the treatment assigned to all rows of the viewr
-#'   object. Currently only able to accept a single treatment per viewr object.
-#' @param stim_param_pos The width of the visual stimulus cycle displayed
-#'   on the positive wall of the chamber.
-#' @param stim_param_neg The width of the visual stimulus cycle displayed
-#'   on the negative wall of the chamber.
-#' @param stim_param_front If using a box-shaped chamber, the width of the
-#'   visual stimulus cycle on the front wall.
+#' object. Currently only able to accept a single treatment per viewr data
+#' object.
 #'
 #' @return A viewr object (tibble or data.frame with attribute
 #'   \code{pathviewR_steps} that includes \code{"treatments added"}). Depending
-#'   on the chamber configuration, it also includes columns for
-#'   \code{vertex_height}, \code{vertex_angle}, \code{pos_wall},
-#'   \code{neg_wall}, \code{front_wall}, and \code{treatment}. This experiment
+#'   on the argument \code{tunnel_config}, the viewr object also includes
+#'   columns storing the values of the supplied arguments. This experimental
 #'   information is also stored in the viewr object's metadata
 #'
 #' @details All length measurements reported in meters.
@@ -2640,11 +2641,7 @@ rm_by_trajnum <- function(obj_name,
 #'
 #' @family utility functions
 #'
-#' @export
-#'
[…]
