### 2021-01-02 83d557d
add flydra example to insert_treatments() documentation
