### 2021-01-02 b5f1445
get_min_dist_v() function with documentation
