+++ R/analytical_functions.R
@@ -572,6 +572,180 @@ Please ensure there are only two columns, ordered x-axis first, y-axis second")
 }
 
 
+#########################         get_min_dist_v        ########################
+
+#' Calculate minimum distance to lateral and end walls V-shaped experimental
+#' tunnel
+#'
+#'
+#' @param obj_name The input viewr object; a tibble or data.frame with attribute
+#'   \code{pathviewR_steps} that includes \code{"viewr"} and
+#'   \code{treatments_added}.
+#' @param simplify If TRUE, the returned object includes only the minimum
+#'   distance between the subject and the lateral/end walls. If FALSE, the
+#'   returned object includes all variables internal to the calculation.
+#'
+#' @return A tibble or data.frame with added variables for
+#'   \code{height_2_vertex}, \code{height_2_screen}, \code{width_2_screen_pos},
+#'   \code{width_2_screen_neg}, \code{min_dist_pos}, \code{min_dist_neg},
+#'   \code{bound_pos}, \code{bound_neg}.
+#'
+#' @details For tunnels in which \code{vertex_angle} is >90˚, \code{bound_pos}
+#' and \code{bound_neg} represent a planes orthogonal to the lateral walls and
+#' are used to modify \code{min_dist_pos} and \code{min_dist_neg} calculations
+#' to prevent erroneous outputs.
+#' \code{calc_min_dist_v()} assumes the subject locomotes facing forward,
+#' therefore \code{min_dist_end} represents the minimum distance between the
+#' subject and the end wall to which it is moving towards
+#' All outputs are in meters.
+#'
+#' @author Eric R. Press
+#'
+#' @family mathematical functions
+#'
+#' @examples
+#'  ## Import sample data from package
+#' motive_data <-
+#'   read_motive_csv(system.file("extdata", "pathviewR_motive_example_data.csv",
+#'                               package = 'pathviewR'))
+#' flydra_data <-
+#'   read_flydra_mat(system.fiule("extdata", pathviewR_motive_example_data.mat",
+#'                               package = 'pathviewR'))
+#'
+#'  ## Clean data up to and including get_full_trajectories()
+#' motive_data_full <-
+#'   motive_data %>%
+#'   relabel_viewr_axes() %>%
+#'   gather_tunnel_data() %>%
+#'   trim_tunnel_outliers() %>%
+#'   rotate_tunnel() %>%
+#'   select_x_percent(desired_percent = 50) %>%
+#'   separate_trajectories(max_frame_gap = "autodetect") %>%
+#'   get_full_trajectories(span = 0.95) %>%
+#'   insert_treatments(tunnel_config = "v",
+#'                    perch_2_vertex = 0.4,
+#'                    vertex_angle = 90,
+#'                    tunnel_length = 2,
+#'                    stim_param_lat_pos = 0.1,
+#'                    stim_param_lat_neg = 0.1,
+#'                    stim_param_end_pos = 0.3,
+#'                    stim_param_end_neg = 0.3,
+#'                    treatment = "lat10_end_30") %>%
+#'
+#'  ## now calculate the minimum distances to each wall
+#'   calc_min_dist_v(simplify = TRUE)
+#'
+#'   ## See 3 new columns for calculations to lateral and end walls
+#'   names(motive_data_full)
+
+calc_min_dist_v <- function(obj_name,
+                            simplify = TRUE){
+
+  ## Check that it's a viewr object
+  if (!any(attr(obj_name,"pathviewR_steps") == "viewr")){
+    stop("This doesn't seem to be a viewr object")
+  }
+
+  ## Check that insert_treatments() has been run
[…]
