### 2021-01-03 6a36758
get_min_dist_box() function and documentation - also fix a few typos in get_min_dist_v()
