+++ R/analytical_functions.R
@@ -572,11 +572,10 @@ Please ensure there are only two columns, ordered x-axis first, y-axis second")
 }
 
 
-#########################         get_min_dist_v        ########################
+#########################         calc_min_dist_v        ########################
 
-#' Calculate minimum distance to lateral and end walls V-shaped experimental
-#' tunnel
-#'
+#' Calculate minimum distance to lateral and end walls in a V-shaped
+#' experimentaltunnel
 #'
 #' @param obj_name The input viewr object; a tibble or data.frame with attribute
 #'   \code{pathviewR_steps} that includes \code{"viewr"} and
@@ -588,7 +587,7 @@ Please ensure there are only two columns, ordered x-axis first, y-axis second")
 #' @return A tibble or data.frame with added variables for
 #'   \code{height_2_vertex}, \code{height_2_screen}, \code{width_2_screen_pos},
 #'   \code{width_2_screen_neg}, \code{min_dist_pos}, \code{min_dist_neg},
-#'   \code{bound_pos}, \code{bound_neg}.
+#'   \code{min_dist_end}, \code{bound_pos}, and \code{bound_neg}.
 #'
 #' @details For tunnels in which \code{vertex_angle} is >90˚, \code{bound_pos}
 #' and \code{bound_neg} represent a planes orthogonal to the lateral walls and
@@ -632,10 +631,10 @@ Please ensure there are only two columns, ordered x-axis first, y-axis second")
 #'                    stim_param_end_neg = 0.3,
 #'                    treatment = "lat10_end_30") %>%
 #'
-#'  ## now calculate the minimum distances to each wall
+#'  ## Now calculate the minimum distances to each wall
 #'   calc_min_dist_v(simplify = TRUE)
 #'
-#'   ## See 3 new columns for calculations to lateral and end walls
+#'   ## See 3 new variables for calculations to lateral and end walls
 #'   names(motive_data_full)
 
 calc_min_dist_v <- function(obj_name,
@@ -742,6 +741,132 @@ calc_min_dist_v <- function(obj_name,
       c(attr(obj_name, "pathviewR_steps"), "min_dist_calculated")
     return(obj_name)
   }
+}
+
+
+#########################       calc_min_dist_box      ##########################
+
+#' Calculate minimum distance to lateral and end walls in a box-shaped
+#' experimental tunnel
+#'
+#' @param obj_name The input viewr object; a tibble or data.frame with attribute
+#'   \code{pathviewR_steps} that includes \code{"viewr"} and
+#'   \code{treatments_added}.
+#'
+#' @return A tibble or data.frame with added variables for
+#'   \code{min_dist_pos}, \code{min_dist_neg}, and \code{min_dist_end},.
+#'
+#' @details \code{calc_min_dist_box()} assumes the subject locomotes facing
+#' forward, therefore \code{min_dist_end} represents the minimum distance
+#' between the subject and the end wall to which it is moving towards.
+#' All outputs are in meters.
+#'
+#' @author Eric R. Press
+#'
+#' @family visual perception functioins
+#'
+#' @examples
+#' ## Import sample data from package
+#'  flydra_data <-
+#'  read_flydra_mat(system.fiule("extdata", pathviewR_motive_example_data.mat",
+#'                               package = 'pathviewR'))
+#'
+#'   ## Clean data up to and including get_full_trajectories()
+#'  flydra_data_full <-
+#'   flydra_data %>%
+#'   redefine_tunnel_center(length_method = "middle",
+#'                         height_method = "user-defined",
[…]
