### 2021-01-03 d5d1e2f
get_vis_angle() function and documentation - other language fixes
