+++ R/analytical_functions.R
@@ -602,7 +602,7 @@ Please ensure there are only two columns, ordered x-axis first, y-axis second")
 #'
 #' @family mathematical functions
 #'
-#' @examples
+#' @example
 #'  ## Import sample data from package
 #' motive_data <-
 #'   read_motive_csv(system.file("extdata", "pathviewR_motive_example_data.csv",
@@ -611,7 +611,7 @@ Please ensure there are only two columns, ordered x-axis first, y-axis second")
 #'   read_flydra_mat(system.fiule("extdata", pathviewR_motive_example_data.mat",
 #'                               package = 'pathviewR'))
 #'
-#'  ## Clean data up to and including get_full_trajectories()
+#'  ## Process data up to and including insert_treatments()
 #' motive_data_full <-
 #'   motive_data %>%
 #'   relabel_viewr_axes() %>%
@@ -750,7 +750,7 @@ calc_min_dist_v <- function(obj_name,
 #' experimental tunnel
 #'
 #' @param obj_name The input viewr object; a tibble or data.frame with attribute
-#'   \code{pathviewR_steps} that includes \code{"viewr"} and
+#'   \code{pathviewR_steps} that include \code{"viewr"} and
 #'   \code{treatments_added}.
 #'
 #' @return A tibble or data.frame with added variables for
@@ -765,13 +765,13 @@ calc_min_dist_v <- function(obj_name,
 #'
 #' @family visual perception functioins
 #'
-#' @examples
+#' @example
 #' ## Import sample data from package
 #'  flydra_data <-
 #'  read_flydra_mat(system.fiule("extdata", pathviewR_motive_example_data.mat",
 #'                               package = 'pathviewR'))
 #'
-#'   ## Clean data up to and including get_full_trajectories()
+#'   ## Process data up to and including insert_treatments()
 #'  flydra_data_full <-
 #'   flydra_data %>%
 #'   redefine_tunnel_center(length_method = "middle",
@@ -830,8 +830,84 @@ calc_min_dist_box <- function(obj_name){
 }
 
 
-## get_vis_angle
-## include vis_angle for end wall
+#########################        get_vis_angle       ###########################
+
+#' Estimate visual angles from a subject's perspective in an experimental tunnel
+#'
+#' @param obj_name The input viewr object; a tibble or data.frame with
+#' attributes \code{pathviewR_steps} that include \code{"viewr"}
+#' and \code{min_dist_calculated}.
+#'
+#' @details \code{get_vis_angle()} assumes the subject's gaze is fixed at the
+#'   point on the either side of the tunnel that minimizes the distance to
+#'   visual stimuli and therefore maximizes visual angles. Angles are reported
+#'   in radians eg. \code{vis_angle_pos_rad} or degrees eg.
+#'   \code{vis_angle_pos_deg}.
+#'
+#' @return A tibble or data.frame with added variables for
+#'   \code{vis_angle_pos_rad}, \code{vis_angle_pos_deg},
+#'   \code{vis_angle_neg_rad}, \code{vos_angle_neg_deg},
+#'   \code{vis_angle_end_rad}, and \code{vis_angle_end_deg}.
+#'
+#' @author Eric R. Press
+#'
+#' @family visual perception functions
+#'
+#' @examples
+#'  ## Import sample data from package
+#' motive_data <-
+#'   read_motive_csv(system.file("extdata", "pathviewR_motive_example_data.csv",
+#'                               package = 'pathviewR'))
+#' flydra_data <-
+#'   read_flydra_mat(system.fiule("extdata", pathviewR_motive_example_data.mat",
+#'                               package = 'pathviewR'))
+#'
[…]
