### 2021-01-03 99c9ef5
get_sf() function with documentation
