+++ R/analytical_functions.R
@@ -841,8 +841,8 @@ calc_min_dist_box <- function(obj_name){
 #' @details \code{get_vis_angle()} assumes the subject's gaze is fixed at the
 #'   point on the either side of the tunnel that minimizes the distance to
 #'   visual stimuli and therefore maximizes visual angles. Angles are reported
-#'   in radians eg. \code{vis_angle_pos_rad} or degrees eg.
-#'   \code{vis_angle_pos_deg}.
+#'   in radians/cycle (\code{vis_angle_pos_rad}) or degrees/cycle
+#'   (\code{vis_angle_pos_deg}).
 #'
 #' @return A tibble or data.frame with added variables for
 #'   \code{vis_angle_pos_rad}, \code{vis_angle_pos_deg},
@@ -952,219 +952,98 @@ get_vis_angle <- function(obj_name){
 #' perspective in an experimental tunnel.
 #'
 #' @param obj_name The input viewr object; a tibble or data.frame with attribute
-#' \code{pathviewR_steps} that includes \code{"viewr"}
-#' @param vertex_angle The angle (in degrees) subtended by a vertical axis and
-#' the sides of the tunnel. Equivalent to the angle of the "V" divided by 2.
-#' \code{vertex_angle)} defaults to 45.
-#' @param stim_param_pos The length (in meters) of the visual stimulus presented
-#' on the positive side of the tunnel (i.e. \code{position_width >= 0}). For
-#' example, a sinusoidal grating 10cm wide is \code{stim_param_pos = 0.1}
-#' @param stim_param_neg The same convention as \code{stim_param_pos} but for
-#' stimuli presented on the negative side of the tunnel
-#' (i.e. \code{position_width < 0}).
-#' @param simplify_output If TRUE, the returned object includes minimum distance
-#' as well as the spatial frequencies perceived on either side of the
-#' tunnel. If FALSE, the returned object includes all variables internal to the
-#' calculation.
+#' \code{pathviewR_steps} that includes \code{"viewr"} and
+#' \code{vis_angles_calculated}.
 #'
-#' @details \code{cal_sf_V} assumes fixed gaze at the point on the
-#' either side of the tunnel that minimizes the distance to visual stimuli and
-#' thereby minimizes spatial frequencies.
+#' @details \code{get_sf} in cycles/degree is the inverse of visual angle
+#' (degrees/cycle).
 #'
 #' @return A tibble or data.frame with added variables for
-#' \code{s_freq_pos} and \code{s_freq_neg} reported in cycles per 1deg of visual
+#' \code{sf_pos}, \code{sf_neg}, and \code{sf_end}.
 #' angle.
 #'
 #' @author Eric R. Press
 #'
 #' @family visual perception functions
 #'
-#' @export
+#' @example
+#'  ## Import sample data from package
+#' motive_data <-
+#'   read_motive_csv(system.file("extdata", "pathviewR_motive_example_data.csv",
+#'                               package = 'pathviewR'))
+#' flydra_data <-
+#'   read_flydra_mat(system.fiule("extdata", pathviewR_motive_example_data.mat",
+#'                               package = 'pathviewR'))
+#'
+#'  ## Process data up to and including get_vis_angle()
+#' motive_data_full <-
+#'   motive_data %>%
+#'   relabel_viewr_axes() %>%
+#'   gather_tunnel_data() %>%
+#'   trim_tunnel_outliers() %>%
+#'   rotate_tunnel() %>%
+#'   select_x_percent(desired_percent = 50) %>%
+#'   separate_trajectories(max_frame_gap = "autodetect") %>%
+#'   get_full_trajectories(span = 0.95) %>%
[…]
