### 2021-01-03 f8b74be
update NAMESPACE because install and restart failed due to new function names
