+++ NAMESPACE
@@ -3,10 +3,8 @@
 export("%>%")
 export(as_viewr)
 export(bind_viewr_objects)
-export(calc_sf_V)
-export(calc_sf_box)
-export(calc_vis_angle_V)
-export(calc_vis_angle_box)
+export(calc_min_dist_v)
+export(calc_min_dist_box)
 export(clean_viewr)
 export(clean_viewr_batch)
 export(deg_2_rad)
@@ -20,7 +18,9 @@ export(get_3d_cross_prod)
 export(get_dist_point_line)
 export(get_full_trajectories)
 export(get_header_viewr)
+export(get_sf)
 export(get_velocity)
+export(get_vis_angle)
 export(import_and_clean_batch)
 export(import_and_clean_viewr)
 export(import_batch)
