+++ man/get_sf.Rd
@@ -1,0 +1,35 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/analytical_functions.R
+\name{get_sf}
+\alias{get_sf}
+\title{Estimate the spatial frequency of visual stimuli from the subject's
+perspective in an experimental tunnel.}
+\usage{
+get_sf(obj_name)
+}
+\arguments{
+\item{obj_name}{The input viewr object; a tibble or data.frame with attribute
+\code{pathviewR_steps} that includes \code{"viewr"} and
+\code{vis_angles_calculated}.}
+}
+\value{
+A tibble or data.frame with added variables for
+\code{sf_pos}, \code{sf_neg}, and \code{sf_end}.
+angle.
+}
+\description{
+Estimate the spatial frequency of visual stimuli from the subject's
+perspective in an experimental tunnel.
+}
+\details{
+\code{get_sf} in cycles/degree is the inverse of visual angle
+(degrees/cycle).
+}
+\seealso{
+Other visual perception functions: 
+\code{\link{get_vis_angle}()}
+}
+\author{
+Eric R. Press
+}
+\concept{visual perception functions}
