+++ man/get_3d_cross_prod.Rd
@@ -24,6 +24,7 @@ get_3d_cross_prod(v1, v2)
 }
 \seealso{
 Other mathematical functions: 
+\code{\link{calc_min_dist_v}()},
 \code{\link{deg_2_rad}()},
 \code{\link{find_curve_elbow}()},
 \code{\link{get_2d_angle}()},
