+++ man/calc_min_dist_v.Rd
@@ -1,0 +1,53 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/analytical_functions.R
+\name{calc_min_dist_v}
+\alias{calc_min_dist_v}
+\title{Calculate minimum distance to lateral and end walls in a V-shaped
+experimentaltunnel}
+\usage{
+calc_min_dist_v(obj_name, simplify = TRUE)
+}
+\arguments{
+\item{obj_name}{The input viewr object; a tibble or data.frame with attribute
+\code{pathviewR_steps} that includes \code{"viewr"} and
+\code{treatments_added}.}
+
+\item{simplify}{If TRUE, the returned object includes only the minimum
+distance between the subject and the lateral/end walls. If FALSE, the
+returned object includes all variables internal to the calculation.}
+}
+\value{
+A tibble or data.frame with added variables for
+\code{height_2_vertex}, \code{height_2_screen}, \code{width_2_screen_pos},
+\code{width_2_screen_neg}, \code{min_dist_pos}, \code{min_dist_neg},
+\code{min_dist_end}, \code{bound_pos}, and \code{bound_neg}.
+}
+\description{
+Calculate minimum distance to lateral and end walls in a V-shaped
+experimentaltunnel
+}
+\details{
+For tunnels in which \code{vertex_angle} is >90˚, \code{bound_pos}
+and \code{bound_neg} represent a planes orthogonal to the lateral walls and
+are used to modify \code{min_dist_pos} and \code{min_dist_neg} calculations
+to prevent erroneous outputs.
+\code{calc_min_dist_v()} assumes the subject locomotes facing forward,
+therefore \code{min_dist_end} represents the minimum distance between the
+subject and the end wall to which it is moving towards
+All outputs are in meters.
+}
+\seealso{
+Other mathematical functions: 
+\code{\link{deg_2_rad}()},
+\code{\link{find_curve_elbow}()},
+\code{\link{get_2d_angle}()},
+\code{\link{get_3d_angle}()},
+\code{\link{get_3d_cross_prod}()},
+\code{\link{get_dist_point_line}()},
+\code{\link{get_velocity}()},
+\code{\link{rad_2_deg}()}
+}
+\author{
+Eric R. Press
+}
+\concept{mathematical functions}
