+++ man/get_velocity.Rd
@@ -85,6 +85,7 @@ motive_data <-
 }
 \seealso{
 Other mathematical functions: 
+\code{\link{calc_min_dist_v}()},
 \code{\link{deg_2_rad}()},
 \code{\link{find_curve_elbow}()},
 \code{\link{get_2d_angle}()},
