+++ man/calc_min_dist_box.Rd
@@ -1,0 +1,32 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/analytical_functions.R
+\name{calc_min_dist_box}
+\alias{calc_min_dist_box}
+\title{Calculate minimum distance to lateral and end walls in a box-shaped
+experimental tunnel}
+\usage{
+calc_min_dist_box(obj_name)
+}
+\arguments{
+\item{obj_name}{The input viewr object; a tibble or data.frame with attribute
+\code{pathviewR_steps} that include \code{"viewr"} and
+\code{treatments_added}.}
+}
+\value{
+A tibble or data.frame with added variables for
+\code{min_dist_pos}, \code{min_dist_neg}, and \code{min_dist_end},.
+}
+\description{
+Calculate minimum distance to lateral and end walls in a box-shaped
+experimental tunnel
+}
+\details{
+\code{calc_min_dist_box()} assumes the subject locomotes facing
+forward, therefore \code{min_dist_end} represents the minimum distance
+between the subject and the end wall to which it is moving towards.
+All outputs are in meters.
+}
+\author{
+Eric R. Press
+}
+\concept{visual perception functioins}
