+++ man/get_dist_point_line.Rd
@@ -44,6 +44,7 @@ get_dist_point_line(
 }
 \seealso{
 Other mathematical functions: 
+\code{\link{calc_min_dist_v}()},
 \code{\link{deg_2_rad}()},
 \code{\link{find_curve_elbow}()},
 \code{\link{get_2d_angle}()},
