### 2021-01-03 50b8217
test new argument check for positive numerical values in insert_treatments()
