### 2021-01-03 429258b
port over new insert_treatments() with argument checks for positive numerical values
