+++ R/utility_functions.R
@@ -2708,11 +2708,10 @@ rm_by_trajnum <- function(obj_name,
 #' names(motive_v)
 #' names(flydra_box)
 
-
 insert_treatments <- function(obj_name,
                               tunnel_config = "box",
                               perch_2_vertex = NULL,
-                              vertex_angle = 90,
+                              vertex_angle = NULL, # actual angle of vertex
                               tunnel_width = NULL,
                               tunnel_length = NULL,
                               stim_param_lat_pos = NULL,
@@ -2731,12 +2730,25 @@ insert_treatments <- function(obj_name,
     stop("Run get_full_trajectories() prior to use")
   }
 
-
   ## Translate arguments into variables at beginning of data frame
-  ##
-  ## figure out how to do it such that only the arguments supplied are added to
-  ## the data frame
   if (tunnel_config == "v"){
+    ## Check that relevant numerical arguments have positive values
+    if (perch_2_vertex < 0){
+      stop("perch_2_vertex must have a positive value.")
+    } else if (vertex_angle < 0){
+      stop("vertex_angle must have a positive value")
+    } else if (tunnel_length < 0){
+      stop("tunnel_length must have a positive value")
+    } else if (stim_param_lat_pos < 0){
+      stop("stim_param_lat_pos must have a positive value")
+    } else if (stim_param_lat_neg < 0){
+      stop("stim_param_lat_neg must have a positive value")
+    } else if (stim_param_end_pos < 0){
+      stop("stim_param_end_pos must have a positive value")
+    } else if (stim_param_end_neg < 0){
+      stop("stim_param_end_neg must have a positive value")
+    }
+
     obj_name <- tibble::add_column(obj_name, .before = "frame",
                                    tunnel_config = tunnel_config,
                                    perch_2_vertex = perch_2_vertex,
@@ -2748,6 +2760,21 @@ insert_treatments <- function(obj_name,
                                    stim_param_end_neg = stim_param_end_neg,
                                    treatment = treatment)
   } else if (tunnel_config == "box"){
+    ## Check that relevant numerical arguments have positive values
+    if (tunnel_width < 0){
+      stop("tunnel_width must have a positive value")
+    } else if (tunnel_length < 0){
+      stop("tunnel_length must have a positive value")
+    } else if (stim_param_lat_pos < 0){
+      stop("stim_param_lat_pos must have a positive value")
+    } else if (stim_param_lat_neg < 0){
+      stop("stim_param_lat_neg must have a positive value")
+    } else if (stim_param_end_pos < 0){
+      stop("stim_param_end_pos must have a positive value")
+    } else if (stim_param_end_neg < 0){
+      stop("stim_param_end_neg must have a positive value")
+    }
+
     obj_name <- tibble::add_column(obj_name, .before = "frame",
                                    tunnel_config = tunnel_config,
                                    tunnel_width = tunnel_width,
@@ -2787,9 +2814,3 @@ insert_treatments <- function(obj_name,
                                          "treatments_added")
   return(obj_name)
 }
-
-
-
-
-
-
