### 2021-01-03 46ba850
explain assumtions in get_vis_angle() and get_sf()
