+++ _pkgdown.yml
@@ -58,10 +58,10 @@ reference:
   contents:
   - find_curve_elbow
   - get_velocity
-  - calc_vis_angle_V
-  - calc_vis_angle_box
-  - calc_sf_V
-  - calc_sf_box
+  - calc_min_dist_box
+  - calc_min_dist_v
+  - get_vis_angle
+  - get_sf
   - get_dist_point_line
   - get_3d_cross_prod
   - rad_2_deg
