+++ example_scripts/ropensci_reviews_checklist.md
@@ -36,9 +36,17 @@ functions (and standardised_tunnel, although there is a reason given in the R
 help – because no example data is provided with pathviewR to run the code on).
 
 Items for us:  
-- [ ] consider and/or make examples for calc_sf_box(), calc_sf_V(),
+- [x] consider and/or make examples for calc_sf_box(), calc_sf_V(),
 calc_vis_angle_box() and calc_vis_angle_V() (OR, make examples for whatever
-visual perception function(s) we make)
+visual perception function(s) we make).
+  
+*These functions have been replaced/renamed but all have examples provided.*
+
+- get_sf() [(99c9ef5)](https://github.com/vbaliga/pathviewR/commit/99c9ef517e265cd27ba790b1cb28929e1986e469)
+- get_vis_angle() [(dfd1e2f)](https://github.com/vbaliga/pathviewR/commit/d5d1e2f1726472505d19c091ed4363590fddc855)
+- get_min_dist_box() [(6a36758)](https://github.com/vbaliga/pathviewR/commit/6a367587a06a30780cc8568eb19935237aa9ccb8)
+- get_min_dist_v() [(b5f1445)](https://github.com/vbaliga/pathviewR/commit/b5f1445b743695ed27acb607321d71b34a6a689a)
+
 
 #### **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with URL, BugReports and Maintainer (which may be autogenerated via Authors@R).
 
@@ -177,10 +185,19 @@ am I right? Is there any reason for this? If the front wall is not relevant,
 maybe there is no need to include the option to add parameters for it in the
 insert_treatments() function?
 
-Items for us:  
-- [ ] consider adding in this functionality, OR state in the Help files for
+Items for us: 
+
+- [x] consider adding in this functionality, OR state in the Help files for
 these functions that it is something that will be developed in future pathviewR
 updates
+
+*The visual perception functions now include calculations for the end walls, though the outputs from these functions only include the end wall to which the subject is moving towards.* 
+
+- get_sf() [(99c9ef5)](https://github.com/vbaliga/pathviewR/commit/99c9ef517e265cd27ba790b1cb28929e1986e469)
+- get_vis_angle() [(dfd1e2f)](https://github.com/vbaliga/pathviewR/commit/d5d1e2f1726472505d19c091ed4363590fddc855)
+- get_min_dist_box() [(6a36758)](https://github.com/vbaliga/pathviewR/commit/6a367587a06a30780cc8568eb19935237aa9ccb8)
+- get_min_dist_v() [(b5f1445)](https://github.com/vbaliga/pathviewR/commit/b5f1445b743695ed27acb607321d71b34a6a689a)
+
 
 *Note from VBB:* for brevity, the details of the reviewer's calculations will
 not be included here.
@@ -189,15 +206,21 @@ not be included here.
 these lines could simply replace these lines and these lines in calc_sf_box()
 and calc_vis_angle_box()
 
-Items for us:  
-- [ ] replace `ifelse()` statement in `calc_sf_box()`   
-- [ ] replace `ifelse()` statement in `calc_vis_angle_box()`
+Items for us: 
+
+- [x] replace `ifelse()` statement in `calc_sf_box()`   
+- [x] replace `ifelse()` statement in `calc_vis_angle_box()`
+
+- get_sf() [(99c9ef5)](https://github.com/vbaliga/pathviewR/commit/99c9ef517e265cd27ba790b1cb28929e1986e469)
+- get_vis_angle() [(dfd1e2f)](https://github.com/vbaliga/pathviewR/commit/d5d1e2f1726472505d19c091ed4363590fddc855)
 
 > Can you please correct the calculations, and adapt the
 test-calc_vis_angle_box.R file?
 
 Items for us:  
[…]
