### 2021-01-03 35480e3
updates to insert_treatments() portion of visual percptions vignette
