### 2021-01-03 64133e5
change simplify argument to simplify_output for calc_min_dist_v
