### 2021-01-03 6225b0c
Add calc_min_dist() and get_vis_angle() examples with visualizations
