### 2021-01-04 a18843c
update roxygen documentation
