### 2021-01-04 a1b0fe0
Add introduction (no figures yet) and reorganize vignette
