+++ vignettes/visual-perception-functions.Rmd
@@ -17,12 +17,16 @@ knitr::opts_chunk$set(
 )
 ```
 
-`pathviewR` includes functions that estimate visual perceptions based on the 
-distance between the subject and visual stimuli on the walls of the 
-experimental tunnel.
+## Introduction
+
+  Studies of visually guided locomotion in birds, insects or even humans during often involve data gathered from motion capture technologies such as Optitrack's [(Motive)](https://optitrack.com/software/motive/), or the Straw Lab's [(Flydra)](https://github.com/strawlab/flydra). For these experiments, it is important to understand how visual stimuli influence behaviour. While is it not possible to measure how subjects directly perceive visual stimuli, it is possible to use the same motion capture data to calculate estimates of stimulus properties as they are perceived by the subject. With the tools available in pathviewR, researchers in ethology or psychology can provide analyses of both stimulus and response under natural locomotor conditions. 
+  *Inherent to estimations of visual perceptions, we make several assumptions, which will be discussed below. We welcome suggestions and aim to address any assumptions that limit the accuracy of these estimates*. 
+  To bridge the gap between objective measures of subject position and estimates of subjective stimulus perception, we can begin by calculating the angle a visual pattern subtends on the subject's retina - the visual angle. 
+  Visual angles can be calculated provided there is information about the physical size of the pattern  and its distance from the subject's retina. Because researchers can control or measure the size of a  pattern, and we can calculate the distance between the subject and pattern using motion capture data, we can further calculate the visual angle produced by any pattern in the visual scene. Therefore, we first need to calculate the distance to the pattern. *Currently, we assume the subject's gaze is directly frontal or lateral to the face, in effect estimating image properties at a point in the frontal and lateral fields of view, respectively*. 
+  
 
 ## Loading packages
-We'll need to start by loading `pathviewR` as well as `tidyverse` for visualizing the data.
+  To begin, let's load `pathviewR` as well as `tidyverse`.
 
 ```{r package_loading, message=FALSE, warning=FALSE}
 library(pathviewR)
@@ -32,26 +36,20 @@ library(magrittr)
 
 
 ## Data preparation
-Data objects must be prepared as described in [the Data Import and Cleaning vignette](https://vbaliga.github.io/pathviewR/articles/data-import-cleaning.html) 
-pipeline prior to their use in these functions. For a detailed description of 
-these functions, please see the linked vignette. 
+Data objects must be prepared as described in [the Data Import and Cleaning vignette](https://vbaliga.github.io/pathviewR/articles/data-import-cleaning.html) pipeline prior to their use with these functions. For a detailed description of the importing and cleaning functions, and when to use them, please see the linked vignette. 
 
-Let's work with a few example datasets included in `pathviewR`. 
-`pathviewR_motive_example_data.csv` is a .csv file exported from `Motive`. 
-`pathviewR_flydra_example_data.mat` is a .mat file exported from `Flydra`.
-`pathviewR` contains an all-in-one cleaning function `clean_viewr` for more 
-coarse-grained data cleaning tasks. We will use this function in the examples 
-below. 
[…]
