### 2021-01-04 8c55c36
add @export to function documentation
