### 2021-01-04 8919178
change @example to @examples in function documentation and fix some code typos in the example sections
