+++ NAMESPACE
@@ -3,6 +3,8 @@
 export("%>%")
 export(as_viewr)
 export(bind_viewr_objects)
+export(calc_min_dist_box)
+export(calc_min_dist_v)
 export(clean_viewr)
 export(clean_viewr_batch)
 export(deg_2_rad)
@@ -16,10 +18,13 @@ export(get_3d_cross_prod)
 export(get_dist_point_line)
 export(get_full_trajectories)
 export(get_header_viewr)
+export(get_sf)
 export(get_velocity)
+export(get_vis_angle)
 export(import_and_clean_batch)
 export(import_and_clean_viewr)
 export(import_batch)
+export(insert_treatments)
 export(plot_by_subject)
 export(plot_viewr_trajectories)
 export(quick_separate_trajectories)

[…]
