### 2021-01-04 fb3fabb
tinkering with visualizations
