### 2021-01-04 baaf82f
fix bugs in @examples for calc_min_dist_box, get_vis_angle and get_sf
