+++ R/analytical_functions.R
@@ -782,10 +782,9 @@ calc_min_dist_v <- function(obj_name,
 #' #   select_x_percent(desired_percent = 50) %>%
 #' #   separate_trajectories(max_frame_gap = "autodetect") %>%
 #' #   get_full_trajectories(span = 0.95) %>%
-#' #   insert_treatments(tunnel_config = "v",
-#' #                    perch_2_vertex = 0.4,
-#' #                    vertex_angle = 90,
-#' #                    tunnel_length = 2,
+#' #   insert_treatments(tunnel_config = "box",
+#' #                    tunnel_length = 3,
+#' #                    tunnel_width = 1,
 #' #                    stim_param_lat_pos = 0.1,
 #' #                    stim_param_lat_neg = 0.1,
 #' #                    stim_param_end_pos = 0.3,
@@ -903,10 +902,9 @@ calc_min_dist_box <- function(obj_name){
 #'   select_x_percent(desired_percent = 50) %>%
 #'   separate_trajectories(max_frame_gap = "autodetect") %>%
 #'   get_full_trajectories(span = 0.95) %>%
-#'   insert_treatments(tunnel_config = "v",
-#'                    perch_2_vertex = 0.4,
-#'                    vertex_angle = 90,
-#'                    tunnel_length = 2,
+#'   insert_treatments(tunnel_config = "box",
+#'                    tunnel_length = 3,
+#'                    tunnel_width = 1,
 #'                    stim_param_lat_pos = 0.1,
 #'                    stim_param_lat_neg = 0.1,
 #'                    stim_param_end_pos = 0.3,
@@ -1027,10 +1025,9 @@ get_vis_angle <- function(obj_name){
 #'   select_x_percent(desired_percent = 50) %>%
 #'   separate_trajectories(max_frame_gap = "autodetect") %>%
 #'   get_full_trajectories(span = 0.95) %>%
-#'   insert_treatments(tunnel_config = "v",
-#'                    perch_2_vertex = 0.4,
-#'                    vertex_angle = 90,
-#'                    tunnel_length = 2,
+#'   insert_treatments(tunnel_config = "box",
+#'                    tunnel_length = 3,
+#'                    tunnel_width = 1,
 #'                    stim_param_lat_pos = 0.1,
 #'                    stim_param_lat_neg = 0.1,
 #'                    stim_param_end_pos = 0.3,
