### 2021-01-04 90552e7
fix error in calc_min_dist_v documentation example
