### 2021-01-04 b873d57
tinkering with the visualizations
