### 2021-01-04 beb84a8
fix insert_treatments example and redocument
hopefully this clears up the last of the "examples" issues. the testthat suite will still throw errors since it hasn't been updated yet
