+++ man/insert_treatments.Rd
@@ -76,7 +76,8 @@ All length measurements reported in meters.
                               package = 'pathviewR'))
  flydra_data <-
  read_flydra_mat(system.file("extdata", "pathviewR_flydra_example_data.mat",
-                              package = 'pathviewR'))
+                              package = 'pathviewR'),
+                              subject_name = "birdie_sanders")
 
   ## Clean data up to and including get_full_trajectories()
 motive_data_full <-
