### 2021-01-05 da4c523
Add images for vignette figure and try to attach
