### 2021-01-07 33ecffe
update image and resize/attach to Rmd with knitr::include_graphics function
