### 2021-01-08 b4afca9
get_sf tests -rename test file from calc_sf_box.R
