### 2021-01-08 6da2639
add motive example tests
