### 2021-01-08 ba8d813
get_vis_angle() tests
