### 2021-01-08 93808ea
calc_min_dist function tests
