### 2021-01-08 a9cc804
tests for insert_treatments
