+++ tests/testthat/test-insert_treaments.R
@@ -1,18 +1,47 @@
-## Import the motive and flydra example datasets
-motive_data <-
-  read_motive_csv(system.file("extdata", "pathviewR_motive_example_data.csv",
-                              package = 'pathviewR'))
+## Import motive data
+motive_data <- # import
+  read_motive_csv(
+    system.file("extdata", "pathviewR_motive_example_data.csv",
+                package = 'pathviewR')
+  )
+## Clean motive data
 motive_full <-
   motive_data %>%
-  clean_viewr(desired_percent = 50,
-              max_frame_gap = "autodetect",
-              span = 0.95)
+  clean_viewr(
+    relabel_viewr_axes = TRUE,
+    gather_tunnel_data = TRUE,
+    trim_tunnel_outliers = TRUE,
+    standardization_option = "rotate_tunnel",
+    select_x_percent = TRUE,
+    desired_percent = 50,
+    rename_viewr_characters = FALSE,
+    separate_trajectories = TRUE,
+    max_frame_gap = "autodetect",
+    get_full_trajectories = TRUE,
+    span = 0.95
+  )
 
+## Prep motive data
+motive_treat <-
+  motive_full %>%
+  insert_treatments(tunnel_config = "v",
+                    perch_2_vertex = 0.3855,
+                    vertex_angle = 90,
+                    tunnel_length = 2,
+                    stim_param_lat_pos = 0.05,
+                    stim_param_lat_neg = 0.05,
+                    stim_param_end_pos = 0.1,
+                    stim_param_end_neg = 0.1,
+                    treatment = "latB")
+
+## Import flydra data
 flydra_data <-
-  read_flydra_mat(system.file("extdata", "pathviewR_flydra_example_data.mat",
+  read_flydra_mat(
+    system.file("extdata", "pathviewR_flydra_example_data.mat",
                 package = 'pathviewR'),
     subject_name = "birdie_wooster")
 
+## Clean flydra data
 flydra_full <-
   flydra_data %>%
   clean_viewr(
@@ -25,53 +54,73 @@ flydra_full <-
     height_zero = 1.44,
     get_velocity = FALSE,
     select_x_percent = TRUE,
+    desired_percent = 60,
     rename_viewr_characters = FALSE,
     separate_trajectories = TRUE,
     get_full_trajectories = TRUE
   )
 
-
-
-## Run insert_treatments() on each
-motive_test <-
-  motive_full %>%
-  insert_treatments(vertex_height  = 0.3855,
-                    vertex_angle   = 45,
-                    stim_param_pos = 0.1,
-                    stim_param_neg = 0.2,
-                    treatment      = "latB")
-
-flydra_test <-
+## Prep flydra data
+flydra_treat <-
   flydra_full %>%
-  insert_treatments(pos_wall = 0.5,
-                    neg_wall = 0.5,
-                    front_wall = 1.0,
-                    stim_param_pos = 0.1,
-                    stim_param_neg = 0.1,
-                    stim_param_front = 0.2,
+  insert_treatments(tunnel_config = "box",
+                    tunnel_width = 1,
+                    tunnel_length = 3,
+                    stim_param_lat_pos = 0.05,
+                    stim_param_lat_neg = 0.05,
+                    stim_param_end_pos = 0.1,
+                    stim_param_end_neg = 0.1,
                     treatment = "latB")
+
+test_that("insert_treatments() fails when nonsense is supplied", {
+  expect_error(insert_treatments("steve"))
+  expect_error(insert_treatments(c("a", "b", "c")))
+  expect_error(insert_treatments())
+  expect_error(insert_treatments(flydra_treat,
+                                 tunnel_config = "box",
+                                 tunnel_width = -1, # negative tunnel width
[…]
