+++ vignettes/visual-perception-functions.Rmd
@@ -27,21 +27,22 @@ knitr::opts_chunk$set(
   
   Visual angles can be calculated provided there is information about the physical size of the pattern  and its distance from the subject's retina. Because researchers can control or measure the size of a  pattern, and we can calculate the distance between the subject and pattern using motion capture data, we can further calculate the visual angle produced by patterns in the visual scene. Therefore, we first need to calculate the distance from the subject's retina to the pattern. 
   
-  *Currently, we assume the subject's gaze is directly frontal or lateral to the face, in effect estimating image properties at single points in the frontal and lateral fields of view, respectively. We currently calculate distances to the center of the subject's head rather than the position of the retina - future versions of pathviewR will include features addressing these limitations*. 
+  *Currently, we assume the subject's gaze is directly frontal or lateral to the face, in effect estimating image properties at single points in the frontal and lateral fields of view, respectively. We currently calculate distances to the center of the subject's head rather than the position of the retina - future versions of pathviewR will include features addressing these limitations, such as including head orientation information and eye position relative to the center of the subject's head*. 
   
   
-```{r, echo=FALSE, out.width="100%", fig.cap="Visual angles can be calculated using the size of a visual pattern (`stim_param`) and the distance to the pattern. Larger patterns at shorter distances produce larger visual angles. For a given distance, gratings produce a constant visual angle from one perspectival orientation while dot fields produce constant visual angles from many orientations."}
+```{r, echo=FALSE, out.width="100%", fig.cap="Visual angles can be calculated using the size of a visual pattern (`stim_param`) and the distance to the pattern. Larger patterns at shorter distances produce larger visual angles. For a given distance to a grating pattern a constant visual angle is produced from a single line of sight while dot fields produce constant visual angles from many lines of sight"}
 knitr::include_graphics("https://github.com/vbaliga/pathviewR/raw/master/images/stim_param_angle.jpeg")
 ```
 
   
 
 ## Loading packages
-  To begin, let's load `pathviewR` as well as `tidyverse`.
+  To begin, let's load `pathviewR` and a few of the packages in `tidyverse`.
 
 ```{r package_loading, message=FALSE, warning=FALSE}
 library(pathviewR)
-library(tidyverse)
+library(ggplot2)
+library(magrittr)
 ```
 
 
@@ -310,6 +311,7 @@ Notice the spatial frequency decreases as the subject approaches the end wall th
 
 
 ### Stay tuned as additional features for image motion estimation are coming soon!
+
 
 
 
