### 2021-01-12 fa7b12e
add references to Dakin et al, 2016 and Frost, 2010.
