+++ example_scripts/ropensci_reviews_checklist.md
@@ -10,20 +10,6 @@ full, but all of the items that require action from us should be included here
 Ultimately, the contents of this file will be transformed into our Response to
 Reviews.
 
-As we address items, we should:  
-- check it off in this file  
-- add in the name of the commit to this file using the first 7
-characters of the commit ID
-
-For example:  
-- [x] consolidate visual perception functions into one function
-[(7f666ce)](https://github.com/vbaliga/pathviewR/commit/7f666ce6e9eee09545ddd9e30d18c0c87ab7eeaf)  
-- [ ] get a million dollars
-
-The first item was addressed and the corresponding commit has been
-linked. Note that the item has been checked off via adding an “x” to “[ ]”.
-
-The second item is...a work in progress...
 
 ## Review from Reviewer 1 (asbonnetlebrun)
 
@@ -42,10 +28,10 @@ visual perception function(s) we make).
   
 *These functions have been replaced/renamed but all have examples provided.*
 
-- get_sf() [(99c9ef5)](https://github.com/vbaliga/pathviewR/commit/99c9ef517e265cd27ba790b1cb28929e1986e469)
-- get_vis_angle() [(dfd1e2f)](https://github.com/vbaliga/pathviewR/commit/d5d1e2f1726472505d19c091ed4363590fddc855)
-- get_min_dist_box() [(6a36758)](https://github.com/vbaliga/pathviewR/commit/6a367587a06a30780cc8568eb19935237aa9ccb8)
-- get_min_dist_v() [(b5f1445)](https://github.com/vbaliga/pathviewR/commit/b5f1445b743695ed27acb607321d71b34a6a689a)
+- get_sf() [(99c9ef5)](https://github.com/vbaliga/pathviewR/commit/99c9ef517e265cd27ba790b1cb28929e1986e469)  
+- get_vis_angle() [(dfd1e2f)](https://github.com/vbaliga/pathviewR/commit/d5d1e2f1726472505d19c091ed4363590fddc855)  
+- get_min_dist_box() [(6a36758)](https://github.com/vbaliga/pathviewR/commit/6a367587a06a30780cc8568eb19935237aa9ccb8)  
+- get_min_dist_v() [(b5f1445)](https://github.com/vbaliga/pathviewR/commit/b5f1445b743695ed27acb607321d71b34a6a689a)  
 
 
 #### **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with URL, BugReports and Maintainer (which may be autogenerated via Authors@R).
@@ -127,7 +113,7 @@ separately? (or maybe that is not relevant...)
 
 Items for us:  
 - [x] we already do this. our example includes it so I made a slight modification 
-to the language that will hopefully make it more clear [(62b63b7)](https://github.com/vbaliga/pathviewR/commit/62b63b7c771036b84540ac9a7a5d259cd66fa07b)
+to the language that will hopefully make it more clear [(62b63b7)](https://github.com/vbaliga/pathviewR/commit/62b63b7c771036b84540ac9a7a5d259cd66fa07b)  
 
 #### Visual perception 
 
@@ -140,7 +126,7 @@ tunnels with visual stimuli consisting of sine wave gratings on the tunnel
 walls?
 
 Items for us:   
-- [x] revise the language of the first couple paragraphs accordingly [(97704bc)](https://github.com/vbaliga/pathviewR/commit/97704bcc46d42257aa113562f39297ec619b3a18)
[…]
