+++ vignettes/visual-perception-functions.Rmd
@@ -31,7 +31,7 @@ knitr::opts_chunk$set(
   
   
 ```{r, echo=FALSE, out.width="100%", fig.cap="Visual angles can be calculated using the size of a visual pattern (`stim_param`) and the distance to the pattern. Larger patterns at shorter distances produce larger visual angles. For a given distance to a grating pattern a constant visual angle is produced from a single line of sight while dot fields produce constant visual angles from many lines of sight"}
-knitr::include_graphics("https://github.com/vbaliga/pathviewR/raw/master/images/stim_param_angle.jpeg")
+knitr::include_graphics("https://raw.githubusercontent.com/vbaliga/pathviewR/master/images/stim_param_angle.jpeg")
 ```
 
   
