+++ pkgdown.yml
@@ -5,8 +5,8 @@ articles:
   data-import-cleaning: data-import-cleaning.html
   managing-frame-gaps: managing-frame-gaps.html
   visual-perception-functions: visual-perception-functions.html
-last_built: 2021-01-14T19:12Z
+last_built: 2021-01-14T19:37Z
 urls:
-  reference: https://vbaliga.github.io/pathviewR//reference
-  article: https://vbaliga.github.io/pathviewR//articles
+  reference: https://vbaliga.github.io/pathviewr//reference
+  article: https://vbaliga.github.io/pathviewr//articles
 
