+++ man/pathviewr-package.Rd
@@ -1,0 +1,44 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/pathviewr.R
+\docType{package}
+\name{pathviewr-package}
+\alias{pathviewr}
+\alias{pathviewr-package}
+\title{pathviewr: Wrangle, Analyze, and Visualize Animal Movement Data}
+\description{
+\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}
+
+Tools to import, clean, and visualize movement data,
+    particularly from motion capture systems such as Optitrack's 
+    Motive, the Straw Lab's Flydra, or from other sources. We provide 
+    functions to remove artifacts, standardize tunnel position and tunnel 
+    axes, select a region of interest, isolate specific trajectories, fill
+    gaps in trajectory data, and calculate 3D and per-axis velocity. For 
+    experiments of visual guidance, we also provide functions that use 
+    subject position to estimate perception of visual stimuli.
+}
+\seealso{
+Useful links:
+\itemize{
+  \item \url{https://github.com/vbaliga/pathviewr}
+  \item Report bugs at \url{https://github.com/vbaliga/pathviewr/issues}
+}
+
+}
+\author{
+\strong{Maintainer}: Vikram B. Baliga \email{vbaliga87@gmail.com} (\href{https://orcid.org/0000-0002-9367-8974}{ORCID})
+
+Authors:
+\itemize{
+  \item Melissa S. Armstrong \email{melissa.armstrong@gmail.com} (\href{https://orcid.org/0000-0002-3059-0094}{ORCID})
+  \item Eric R. Press \email{epress12@gmail.com} (\href{https://orcid.org/0000-0002-1944-3755}{ORCID})
+}
+
+Other contributors:
+\itemize{
+  \item Anne-Sophie Bonnet-Lebrun [reviewer]
+  \item Marco Sciaini [reviewer]
+}
+
+}
+\keyword{internal}
