+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   data-import-cleaning: data-import-cleaning.html
   managing-frame-gaps: managing-frame-gaps.html
   visual-perception-functions: visual-perception-functions.html
-last_built: 2021-01-24T22:27Z
+last_built: 2021-01-24T23:58Z
 urls:
   reference: https://vbaliga.github.io/pathviewr//reference
   article: https://vbaliga.github.io/pathviewr//articles
