+++ vignettes/visual-perception-functions.Rmd
@@ -116,7 +116,7 @@ flydra_full <-
 Now that our objects have been cleaned, we must use `insert_treatments()` to add information about the experiments, including relevant properties of the visual stimulus and experimental tunnel that are necessary for calculating visual perceptions.
 *`pathviewr` currently supports rectangular (box) or v-shaped experimental tunnels, though we are intrigued to include additional tunnel configurations*.
 
-If you would like to request any additional features such as other tunnel configurations, please create and issue [(here)](https://github.com/vbaliga/pathviewr/issues/new/choose)
+If you would like to request any additional features such as other tunnel configurations, please create an issue [(here)](https://github.com/vbaliga/pathviewr/issues/new/choose)
 
 #### V-shaped tunnel example
 The data within `motive_full` were collected from birds flying through a 3m long v-shaped tunnel in which the height of the origin `(0,0,0)` was set to the height of the perches that were 0.3855m above the vertex, which was angled at 90˚. The visual stimuli on the positive and negative walls of the tunnel (where `position_width` values > 0 and < 0, respectively) were stationary dot-fields. Each dot was of size 0.05m in diameter. The visual stimuli on the positive and negative end walls of the tunnel (where `position_length` > 0 and < 0, respectively) were dot fields with dots 0.1m in diameter. This treatment was referred to as `"latB"`. 
