### 2021-01-25 f3d20f6
clarify some points related to visual perception commits.
