+++ example_scripts/ropensci_reviews_checklist.md
@@ -32,7 +32,7 @@ functions referred to in this reviewer comment have now been replaced or
 renamed.  
 
 - get_sf() [(99c9ef5)](https://github.com/vbaliga/pathviewr/commit/99c9ef517e265cd27ba790b1cb28929e1986e469)  
-- get_vis_angle() [(dfd1e2f)](https://github.com/vbaliga/pathviewr/commit/d5d1e2f1726472505d19c091ed4363590fddc855)  
+- get_vis_angle() [(dfd1e2f)](https://github.com/vbaliga/pathviewr/commit/d5d1e2f1726472505d19c091ed4363590fddc855)[(6a36758)](https://github.com/vbaliga/pathviewr/commit/6a367587a06a30780cc8568eb19935237aa9ccb8)  
 - get_min_dist_box() [(6a36758)](https://github.com/vbaliga/pathviewr/commit/6a367587a06a30780cc8568eb19935237aa9ccb8)  
 - get_min_dist_v() [(b5f1445)](https://github.com/vbaliga/pathviewr/commit/b5f1445b743695ed27acb607321d71b34a6a689a)  
 
@@ -127,7 +127,7 @@ tunnels with visual stimuli consisting of sine wave gratings on the tunnel
 walls?
 
 Thanks. We have heavily revised the language of the first couple paragraphs 
-accordingly, and a figure has aslo been inserted to hopefully give readers a 
+accordingly, and a figure has also been inserted to hopefully give readers a 
 better sense of the concepts at hand. Please let us know what you think!  [(97704bc)](https://github.com/vbaliga/pathviewr/commit/97704bcc46d42257aa113562f39297ec619b3a18)  
 
 > Also, my understanding is that here you implement only two cases:
@@ -178,7 +178,7 @@ include the end wall towards which the subject is moving towards. Please see
 the changes to following functions:  
 
 - get_sf() [(99c9ef5)](https://github.com/vbaliga/pathviewr/commit/99c9ef517e265cd27ba790b1cb28929e1986e469)  
-- get_vis_angle() [(dfd1e2f)](https://github.com/vbaliga/pathviewr/commit/d5d1e2f1726472505d19c091ed4363590fddc855)  
+- get_vis_angle() [(dfd1e2f)](https://github.com/vbaliga/pathviewr/commit/d5d1e2f1726472505d19c091ed4363590fddc855)[(6a36758)](https://github.com/vbaliga/pathviewr/commit/6a367587a06a30780cc8568eb19935237aa9ccb8)   
 - get_min_dist_box() [(6a36758)](https://github.com/vbaliga/pathviewr/commit/6a367587a06a30780cc8568eb19935237aa9ccb8)  
 - get_min_dist_v() [(b5f1445)](https://github.com/vbaliga/pathviewr/commit/b5f1445b743695ed27acb607321d71b34a6a689a)  
 
@@ -194,15 +194,15 @@ Thanks! We have now replaced the `ifelse()` lines in the corresponding
 functions, which also have been renamed:  
 
 - get_sf() [(99c9ef5)](https://github.com/vbaliga/pathviewr/commit/99c9ef517e265cd27ba790b1cb28929e1986e469)  
-- get_vis_angle() [(dfd1e2f)](https://github.com/vbaliga/pathviewr/commit/d5d1e2f1726472505d19c091ed4363590fddc855)  
+- get_vis_angle()[(6a36758)](https://github.com/vbaliga/pathviewr/commit/6a367587a06a30780cc8568eb19935237aa9ccb8)   
 
 > Can you please correct the calculations, and adapt the
 test-calc_vis_angle_box.R file?
 
 Thanks sincerely for catching this! We have made corrections to the calculations
[…]
