+++ man/pathviewr-package.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/pathviewR.R
+% Please edit documentation in R/pathviewr.R
 \docType{package}
 \name{pathviewr-package}
 \alias{pathviewr}
