### 2021-01-28 81f2a34
minor text edits and bump version
