+++ vignettes/managing-frame-gaps.Rmd
@@ -22,9 +22,10 @@ knitr::opts_chunk$set(
 `pathviewr` defines trajectories as continuous movement from one side of a
 region of interest to the other. Before trajectories can be defined, the region
 of interest must be selected via `select_x_percent()` in addition to all the
-previous steps of the data import and cleaning pipeline as described in [the Data Import and Cleaning vignette](https://vbaliga.github.io/pathviewr/articles/data-import-cleaning.html).
-We'll start with loading `pathviewr` and a few packages from the `tidyverse`
-and importing our data.
+previous steps of the data import and cleaning pipeline as described in 
+[the Data Import and Cleaning vignette](https://vbaliga.github.io/pathviewr/articles/data-import-cleaning.html).
+We'll start with loading `pathviewr` and a few packages from the `tidyverse` and
+importing our data.
 
 ```{r package_loading, message=FALSE, warning=FALSE}
 ## If you do not already have pathviewr installed:
@@ -72,7 +73,7 @@ plot(motive_cleaned$position_length,
 We've cleaned up our data and are ready to link individual data points into
 continuous trajectories. Deciding exactly how to define a trajectory will depend
 largely on the quality of data collected and the resolution required to answer a
-given question about the data. If the data is fairly continuous and high
+given question about the data. If the data are fairly continuous and high
 resolution is required, the default `max_frame_gap = 1` will not allow any frame
 gaps--movement must be continuous.
 
@@ -163,21 +164,26 @@ motive_cleaned %>%
 ```
 
 The output of `visualize_frame_gap_choice()` is a tibble and plot of the number
-of trajectories after running `separate_trajectories()` with `max_frame_gap =
-1`, `max_frame_gap = 2`, `max_frame_gap = 3`, etc. We can see that as the frame
-gap allowance increases, more bits of data are being linked into continuous
-trajectories and thus the total number of trajectories decreases. The vertical
-line on the plot indicates the "elbow" of the plot or the point at which counts
-of trajectories appear to stabilize and increases in the `max_frame_gap`
-allowance no longer effect the trajectory count very much.
+of trajectories after running `separate_trajectories()` with 
+`max_frame_gap = 1`, `max_frame_gap = 2`, `max_frame_gap = 3`, etc. 
+We can see that as the frame gap allowance increases, more bits of data are
+being linked into continuous trajectories and thus the total number of
+trajectories decreases. The vertical line on the plot indicates the "elbow" of
+the plot or the point at which counts of trajectories appear to stabilize and
+increases in the `max_frame_gap` allowance no longer effect the trajectory count
+very much.
 
 ### Autodetect
 Setting `max_frame_gap = "autodetect"` rather than a numeric value uses the
 "elbow" of the plots from `visualize_frame_gap_choice()` to guesstimate the best
-value(s) for `max_frame_gap`. In addition to automatically selecting the
-`max_frame_gap` depending on the data, "autodetect" does so on a per subject
-basis rather than applying the same allowable frame gap to all data in the data
-set since frame gaps can vary between subjects. The cap on how high the
+value(s) for `max_frame_gap`. 
+
+**In addition to automatically selecting the `max_frame_gap` depending on the
+data, `autodetect` does so on a per subject basis rather than applying the same
+allowable frame gap to all data in the data set since frame gaps can vary
+between subjects.** 
+
+The cap on how high the
 `max_frame_gap` can go is defined as a proportion of the capture frame rate and
 set by the `frame_rate_proportion` argument, which defaults to `.10`.
 
