+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: pathviewr
 Title: Wrangle, Analyze, and Visualize Animal Movement Data
-Version: 0.9.4.9000
+Version: 0.9.5
 Authors@R: 
     c(person(given = "Melissa S.", 
              family = "Armstrong",

[…]
