### 2021-01-28 00c25b8
update citation in readme
