### 2021-02-11 #409 @maelle
@asbonnetlebrun @marcosci Friendly reminder :smile_cat: Could you please indicate whether you are happy with the authors' response? Template devguide.ropensci.org/approval2template.html
