### 2021-03-12 #409 @vbaliga
Hi everyone - thanks so much for reviewing `pathviewr` and for all your feedback. We're excited to have this be part of the ropensci org. I have now begun the transfer of the repo and post-review updating

Cheers,
Vikram
🐢
