### 2021-03-15 #409 @vbaliga
Hi @maelle, thanks again for all your help. We have taken care of all the items on the to-do list.
> To-dos:

>       * deactivate the automatic deployment you might have set up
>       * remove styling tweaks from your pkgdown config but keep that config file
>       * replace the whole current `pkgdown` website with a [redirecting page](https://devguide.ropensci.org/redirect.html)
>       * replace your package docs URL with `https://docs.ropensci.org/package_name`
>       * In addition, in your DESCRIPTION file, include the docs link in the `URL` field alongside the link to the GitHub repository, e.g.: `URL: https://docs.ropensci.org/foobar (website) https://github.com/ropensci/foobar`
   * [x]  Fix any links in badges for CI and coverage to point to the ropensci URL. We no longer transfer Appveyor projects to ropensci Appveyor account so after transfer of your repo to rOpenSci's "ropensci" GitHub organization the badge should be `[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/pkgname?branch=master&svg=true)](https://ci.appveyor.com/project/individualaccount/pkgname)`. If Appveyor does not pick up new commits after transfer, you might need to delete and re-create the Appveyor project. (Repo transfers are smoother with GitHub Actions)
   * [x]  Please check you updated the package version post-review version updated and that you documented all changes in NEWS.md

> Should you want to acknowledge your reviewers in your package DESCRIPTION, you can do so by making them `"rev"`-type contributors in the `Authors@R` field (with their consent). More info on this [here](https://devguide.ropensci.org/building.html#authorship).

Happy to! Thanks again for the reviews.

> Welcome aboard! We'd love to host a post about your package - either a [short introduction to it with an example](https://ropensci.org/tech-notes/) for a technical audience or [a longer post with some narrative about its development or something you learned, and an example of its use](https://ropensci.org/blog/) for a broader readership. If you are interested, consult [the blog guide](https://blogguide.ropensci.org/), and tag @stefaniebutland in your reply. She will get in touch about timing and can answer any questions.

Thanks, we are discussing the idea of this and will be sure to reach out to Stefanie if we opt to contribute
