### 2021-03-16 #409 @maelle
Awesome, thanks for the update! Please also tell me if other authors need to be invited to the ropensci GitHub organization.
